% Generated by tools/generate-man.R -- do not edit by hand
\name{as.PlotDate.yyyymmdd}
\alias{as.PlotDate.yyyymmdd}
\title{Convert YYYYMMDD to Plot Date}
\description{
Converts integer date to Date for plotting.
}
\usage{
as.PlotDate.yyyymmdd(yyyymmdd)
}
\arguments{
  \item{yyyymmdd}{Integer date. }
}
\value{
Date object. 
}
\seealso{

Other plotting: \code{\link{plot.boxtext}}, \code{\link{colorblind}}, \code{\link{plot.contour}}, \code{\link{plot.ecdf.lineplot}}, \code{\link{plot.hist2d}}, \code{\link{hline}}, \code{\link{iaw.boxplot}}, \code{\link{plot.labelcurve}}, \code{\link{lambda2color}}, \code{\link{plot.loess.lineplot}}, \code{\link{MixColor}}, \code{\link{plot.native.slope}}, \code{\link{plot.p.arrows}}, \code{\link{plot.pdf.embedfonts}}, \code{\link{pdf.end}}, \code{\link{plot.pdf.start.bera}}, \code{\link{pdf.start}}, \code{\link{plot.timearea}}, \code{\link{plot.worldmap}}, \code{\link{plot.points.toomany}}, \code{\link{plot.shade.xyy}}, \code{\link{plot.text.bg}}, \code{\link{plot.text.onbg}}, \code{\link{vline}}, \code{\link{plot.yearly.tickmarks}}, \code{\link{plot.ylab.left}}
}
