% Generated by tools/generate-man.R -- do not edit by hand
\name{colSds}
\alias{colSds}
\title{Column Standard Deviations}
\description{
Calculates SD for each column.
}
\usage{
colSds(x, na.rm)
}
\arguments{
  \item{x}{Matrix or data frame.}
  \item{na.rm}{Remove NA values. }
}
\value{
Vector of standard deviations. 
}
\seealso{

Other statistics: \code{\link{autocorrel}}, \code{\link{autopcorrel}}, \code{\link{cmpsum}}, \code{\link{colN}}, \code{\link{covp}}, \code{\link{dependent.rank}}, \code{\link{print.cor}}, \code{\link{rank}}, \code{\link{rowSds}}, \code{\link{sdp}}, \code{\link{sse}}, \code{\link{sst}}, \code{\link{summary}}, \code{\link{tabularsummary}}, \code{\link{varp}}
}
\examples{
\dontrun{
# 5x4 matrix: one SD per column
set.seed(1)
m <- matrix(rnorm(20), nrow = 5, ncol = 4)
iaw$colSds(m)

# Data frame of numeric columns
df <- data.frame(a = c(1, 2, 3, 4, 5), b = c(2, 2, 2, 2, 2))
iaw$colSds(df)   # b has SD = 0

# NA values are ignored by default (na.rm = TRUE)
df2 <- data.frame(x = c(1, NA, 3), y = c(4, 5, 6))
iaw$colSds(df2)
}
}
