% Generated by tools/generate-man.R -- do not edit by hand
\name{cumulateseries}
\alias{cumulateseries}
\title{Cumulative Sum of a Time Series}
\description{
Wrapper around cumsum with validation.
}
\usage{
cumulateseries(seriesin)
}
\arguments{
  \item{seriesin}{Numeric vector. }
}
\value{
Numeric vector of cumulative sums. 
}
\seealso{

Other time-series: \code{\link{chgseries}}, \code{\link{compoundseries}}, \code{\link{lagdataframe}}, \code{\link{lagseries}}, \code{\link{leadlagpanel}}, \code{\link{leadlagseries}}, \code{\link{leadseries}}, \code{\link{ma}}, \code{\link{pchgseries}}
}
\examples{
\dontrun{
iaw$cumulateseries(c(1, 2, 3, 4, 5))
}
}
