% Generated by tools/generate-man.R -- do not edit by hand
\name{diffid.for.fixed.effects}
\alias{diffid.for.fixed.effects}
\title{Demean for Fixed Effects}
\description{
Subtracts group means from each observation for use in
fixed-effects regressions. Handles vectors and matrices/data frames.
}
\usage{
diffid.for.fixed.effects(h, id)
}
\arguments{
  \item{h}{Variable(s) to demean (vector, matrix, or data frame).}
  \item{id}{Group identifier (will be coerced to factor). }
}
\value{
De-meaned values with the same dimensions as h. 
}
\seealso{

Other data-manipulation: \code{\link{kill.useless.cols}}, \code{\link{recode}}, \code{\link{rename.columns}}, \code{\link{subset.data.frame}}
}
\examples{
\dontrun{
newy <- iaw$diffid.for.fixed.effects(y, id)
newx <- iaw$diffid.for.fixed.effects(X, id)
iaw$olm(newy ~ newx)
}
}
