% Generated by tools/generate-man.R -- do not edit by hand
\name{enable.error.line.num}
\alias{enable.error.line.num}
\title{Enable Error Line Numbers}
\description{
Enables line numbers in error messages.
}
\usage{
enable.error.line.num()
}
\value{
Invisible NULL. 
}
\seealso{

Other utilities: \code{\link{ascii}}, \code{\link{base}}, \code{\link{beep}}, \code{\link{bisection}}, \code{\link{cat.both}}, \code{\link{cat.stderr}}, \code{\link{chatter}}, \code{\link{check.names}}, \code{\link{debug.advice}}, \code{\link{debug.on}}, \code{\link{dict.lookup}}, \code{\link{done}}, \code{\link{exit}}, \code{\link{grepcolname}}, \code{\link{heartbeat}}, \code{\link{index.of.variable}}, \code{\link{instrumentR}}, \code{\link{kable}}, \code{\link{latex}}, \code{\link{latlon.distance}}, \code{\link{latlonx}}, \code{\link{mklatlonid}}, \code{\link{invlatlon}}, \code{\link{testlatlon}}, \code{\link{ls.objects}}, \code{\link{meminfo}}, \code{\link{mk.random.walk}}, \code{\link{mkRdictionary}}, \code{\link{msg}}, \code{\link{nearest}}, \code{\link{now}}, \code{\link{numobs.by}}, \code{\link{object.size.MB}}, \code{\link{osinfo}}, \code{\link{rdraw.perfect}}, \code{\link{require.variables}}, \code{\link{rexp.perfect}}, \code{\link{rnorm.perfect}}, \code{\link{runall}}, \code{\link{setupknitr}}, \code{\link{sink}}, \code{\link{sprint.data.frame}}, \code{\link{strcat}}, \code{\link{whatis}}, \code{\link{which.nearest}}
}
\examples{
\dontrun{
# Turn on line-number reporting in error messages
iaw$enable.error.line.num()

# Now errors show the source line where they occurred
f <- function() stop("something went wrong")
f()
}

\dontrun{
# Useful at the top of a batch script for easier post-mortem debugging
iaw$enable.error.line.num()
source("my_analysis.R")  # errors now include line numbers
}

\dontrun{
# Check the current setting, enable if not already on
getOption("show.error.locations")  # NULL or FALSE by default
iaw$enable.error.line.num()
getOption("show.error.locations")  # TRUE
}

}
