% Generated by tools/generate-man.R -- do not edit by hand
\name{epoch2nyc}
\alias{epoch2nyc}
\title{Convert Unix Epoch to NYC Time}
\description{
Converts Unix timestamp to NYC timezone.
}
\usage{
epoch2nyc(epoch)
}
\arguments{
  \item{epoch}{Numeric Unix timestamp. }
}
\value{
POSIXct in America/New_York. 
}
\seealso{

Other datetime: \code{\link{diffdays}}, \code{\link{epoch2POSIXct}}, \code{\link{epoch2yyyymmdd}}, \code{\link{tmdiffsec}}, \code{\link{yyyymmdd.toggle}}, \code{\link{yyyymmdd}}
}
