% Generated by tools/generate-man.R -- do not edit by hand
\name{epoch2POSIXct}
\alias{epoch2POSIXct}
\title{Convert Unix Epoch to POSIXct}
\description{
Converts Unix timestamp to POSIXct.
}
\usage{
epoch2POSIXct(epoch, tz = "UTC")
}
\arguments{
  \item{epoch}{Numeric Unix timestamp.}
  \item{tz}{Timezone. Default "UTC". }
}
\value{
POSIXct datetime. 
}
\seealso{

Other datetime: \code{\link{diffdays}}, \code{\link{epoch2nyc}}, \code{\link{epoch2yyyymmdd}}, \code{\link{tmdiffsec}}, \code{\link{yyyymmdd.toggle}}, \code{\link{yyyymmdd}}
}
\examples{
\dontrun{
iaw$epoch2POSIXct(1609459200)
}
}
