% Generated by tools/generate-man.R -- do not edit by hand
\name{epoch2yyyymmdd}
\alias{epoch2yyyymmdd}
\alias{epoch2hhmmss}
\title{Convert Unix Epoch to YYYYMMDD Integer}
\description{
Converts Unix timestamp to integer date.
}
\usage{
epoch2yyyymmdd(epoch, tz = "UTC")
epoch2hhmmss(epoch, tz = "UTC")
}
\arguments{
  \item{epoch}{Numeric Unix timestamp.}
  \item{tz}{Timezone. Default "UTC". }
}
\value{
Integer YYYYMMDD. 
}
\seealso{

Other datetime: \code{\link{diffdays}}, \code{\link{epoch2nyc}}, \code{\link{epoch2POSIXct}}, \code{\link{tmdiffsec}}, \code{\link{yyyymmdd.toggle}}, \code{\link{yyyymmdd}}
}
\examples{
\dontrun{
iaw$epoch2yyyymmdd(1609459200)
}
}
