% Generated by tools/generate-man.R -- do not edit by hand
\name{hline}
\alias{hline}
\title{Add Horizontal Lines to Plot}
\description{
Draws horizontal lines at specified y values.
}
\usage{
hline(atyloc, xrange = NULL, ...)
}
\arguments{
  \item{atyloc}{Y coordinates for lines.}
  \item{xrange}{Optional x range.}
  \item{...}{Graphics parameters. }
}
\value{
Invisible NULL. 
}
\seealso{

Other plotting: \code{\link{as.PlotDate.yyyymmdd}}, \code{\link{plot.boxtext}}, \code{\link{colorblind}}, \code{\link{plot.contour}}, \code{\link{plot.ecdf.lineplot}}, \code{\link{plot.hist2d}}, \code{\link{iaw.boxplot}}, \code{\link{plot.labelcurve}}, \code{\link{lambda2color}}, \code{\link{plot.loess.lineplot}}, \code{\link{MixColor}}, \code{\link{plot.native.slope}}, \code{\link{plot.p.arrows}}, \code{\link{plot.pdf.embedfonts}}, \code{\link{pdf.end}}, \code{\link{plot.pdf.start.bera}}, \code{\link{pdf.start}}, \code{\link{plot.timearea}}, \code{\link{plot.worldmap}}, \code{\link{plot.points.toomany}}, \code{\link{plot.shade.xyy}}, \code{\link{plot.text.bg}}, \code{\link{plot.text.onbg}}, \code{\link{vline}}, \code{\link{plot.yearly.tickmarks}}, \code{\link{plot.ylab.left}}
}
\examples{
\dontrun{
plot(1:10)
iaw$hline(5, col = "red")
}
}
