% Generated by tools/generate-man.R -- do not edit by hand
\name{iaw.boxplot}
\alias{iaw.boxplot}
\alias{boxplot}
\title{Enhanced Boxplot}
\description{
Boxplot with additional features.
}
\usage{
boxplot(x, ...)
}
\arguments{
  \item{x}{Data for boxplot.}
  \item{...}{Boxplot arguments. }
}
\value{
Boxplot object. 
}
\seealso{

Other plotting: \code{\link{as.PlotDate.yyyymmdd}}, \code{\link{plot.boxtext}}, \code{\link{colorblind}}, \code{\link{plot.contour}}, \code{\link{plot.ecdf.lineplot}}, \code{\link{plot.hist2d}}, \code{\link{hline}}, \code{\link{plot.labelcurve}}, \code{\link{lambda2color}}, \code{\link{plot.loess.lineplot}}, \code{\link{MixColor}}, \code{\link{plot.native.slope}}, \code{\link{plot.p.arrows}}, \code{\link{plot.pdf.embedfonts}}, \code{\link{pdf.end}}, \code{\link{plot.pdf.start.bera}}, \code{\link{pdf.start}}, \code{\link{plot.timearea}}, \code{\link{plot.worldmap}}, \code{\link{plot.points.toomany}}, \code{\link{plot.shade.xyy}}, \code{\link{plot.text.bg}}, \code{\link{plot.text.onbg}}, \code{\link{vline}}, \code{\link{plot.yearly.tickmarks}}, \code{\link{plot.ylab.left}}
}
\examples{
\dontrun{
# Note: iaw$boxplot() is a placeholder; use base boxplot() for now
iaw$boxplot(rnorm(100), main = "Example boxplot")

# Multiple groups
iaw$boxplot(list(A = rnorm(50), B = rnorm(50, mean = 1)),
            names = c("Group A", "Group B"))
}

}
