% Generated by tools/generate-man.R -- do not edit by hand
\name{is.numeric}
\alias{is.numeric}
\title{Test if Object is Numeric Vector of Specific Length}
\description{
Extends base is.numeric to also check length.
}
\usage{
is.numeric(nvec, required.length.of.nvec = NULL)
}
\arguments{
  \item{nvec}{Object to test.}
  \item{required.length.of.nvec}{Required length, or \code{NULL} to skip length check (default \code{NULL}). }
}
\value{
Logical scalar. 
}
\seealso{

Other type-checking: \code{\link{is.character}}, \code{\link{is.instring}}, \code{\link{is.scalar}}
}
\examples{
\dontrun{
iaw$is.numeric(5, 1)
iaw$is.numeric(c(1, 2, 3), 3)
}
}
