% Generated by tools/generate-man.R -- do not edit by hand
\name{is.scalar}
\alias{is.scalar}
\title{Test if Object is a Scalar}
\description{
Checks whether an object is a single atomic value.
}
\usage{
is.scalar(x)
}
\arguments{
  \item{x}{Object to test. }
}
\value{
Logical scalar. 
}
\seealso{

Other type-checking: \code{\link{is.character}}, \code{\link{is.instring}}, \code{\link{is.numeric}}
}
\examples{
\dontrun{
iaw$is.scalar(5)
iaw$is.scalar(c(1, 2))
}
}
