% Generated by tools/generate-man.R -- do not edit by hand
\name{kill.useless.cols}
\alias{kill.useless.cols}
\title{Remove Constant Columns}
\description{
Removes columns with no variance.
}
\usage{
kill.useless.cols(d)
}
\arguments{
  \item{d}{Data frame. }
}
\value{
Data frame with constant columns removed. 
}
\seealso{

Other data-manipulation: \code{\link{diffid.for.fixed.effects}}, \code{\link{doublesort}}, \code{\link{recode}}, \code{\link{rename.columns}}, \code{\link{subset.data.frame}}
}
\examples{
\dontrun{
# Column b is constant: removed; a is kept
df <- data.frame(a = 1:3, b = c(1, 1, 1))
iaw$kill.useless.cols(df)

# Multiple constant columns removed at once
df2 <- data.frame(x = rnorm(5), flag = TRUE, id = 99L, y = rnorm(5))
iaw$kill.useless.cols(df2)   # only x and y survive

# All-NA column is treated as constant (one unique value: NA)
df3 <- data.frame(a = c(1, 2, 3), b = c(NA, NA, NA))
iaw$kill.useless.cols(df3)   # b removed
}
}
