% Generated by tools/generate-man.R -- do not edit by hand
\name{lagdataframe}
\alias{lagdataframe}
\title{Lag Data Frame Columns}
\description{
Creates lagged versions of data frame columns.
}
\usage{
lagdataframe(df, vars, nlags = 1, panelid = NULL, timeid = NULL)
}
\arguments{
  \item{df}{Data frame.}
  \item{vars}{Variables to lag.}
  \item{nlags}{Number of lags.}
  \item{panelid}{Panel identifier.}
  \item{timeid}{Time identifier. }
}
\value{
Data frame with lagged columns. 
}
\seealso{

Other time-series: \code{\link{chgseries}}, \code{\link{compoundseries}}, \code{\link{cumulateseries}}, \code{\link{lagseries}}, \code{\link{leadlagpanel}}, \code{\link{leadlagseries}}, \code{\link{leadseries}}, \code{\link{ma}}, \code{\link{pchgseries}}
}
\examples{
\dontrun{
df <- data.frame(t = 1:5, ret = c(0.01, 0.02, -0.01, 0.03, 0.00))

# Add a one-period lag of "ret" (new column "ret.L1")
iaw$lagdataframe(df, vars = "ret")

# Add a two-period lag
iaw$lagdataframe(df, vars = "ret", nlags = 2)

# Lag multiple columns at once
df2 <- data.frame(t = 1:4, ret = 1:4, vol = 10:13)
iaw$lagdataframe(df2, vars = c("ret", "vol"))

}
}
