% Generated by tools/generate-man.R -- do not edit by hand
\name{lambda2color}
\alias{lambda2color}
\title{Wavelength to Color}
\description{
Converts wavelength to RGB color.
}
\usage{
lambda2color(lambda)
}
\arguments{
  \item{lambda}{Wavelength in nm. }
}
\value{
Color value. 
}
\seealso{

Other plotting: \code{\link{as.PlotDate.yyyymmdd}}, \code{\link{plot.boxtext}}, \code{\link{colorblind}}, \code{\link{plot.contour}}, \code{\link{plot.ecdf.lineplot}}, \code{\link{plot.hist2d}}, \code{\link{hline}}, \code{\link{iaw.boxplot}}, \code{\link{plot.labelcurve}}, \code{\link{plot.loess.lineplot}}, \code{\link{MixColor}}, \code{\link{plot.native.slope}}, \code{\link{plot.p.arrows}}, \code{\link{plot.pdf.embedfonts}}, \code{\link{pdf.end}}, \code{\link{plot.pdf.start.bera}}, \code{\link{pdf.start}}, \code{\link{plot.timearea}}, \code{\link{plot.worldmap}}, \code{\link{plot.points.toomany}}, \code{\link{plot.shade.xyy}}, \code{\link{plot.text.bg}}, \code{\link{plot.text.onbg}}, \code{\link{vline}}, \code{\link{plot.yearly.tickmarks}}, \code{\link{plot.ylab.left}}
}
\examples{
# Red light (~700 nm)
iaw$lambda2color(700)

# Blue light (~450 nm)
iaw$lambda2color(450)

# Green light (~550 nm)
iaw$lambda2color(550)

# Boundary of visible spectrum (violet)
iaw$lambda2color(380)

# Vectorise over a small set of wavelengths
sapply(c(400, 500, 600, 700), iaw$lambda2color)

\dontrun{
# Use wavelength colors in a spectrum plot
lambdas <- seq(380, 780, by = 10)
cols <- sapply(lambdas, iaw$lambda2color)
barplot(rep(1, length(lambdas)), col = cols, border = NA,
        names.arg = lambdas, las = 2, main = "Visible spectrum")

# Color-code data points by a wavelength-mapped variable
x <- seq(380, 780, length.out = 50)
y <- sin(seq(0, 4 * pi, length.out = 50))
plot(x, y, pch = 16, col = sapply(x, iaw$lambda2color),
     xlab = "Wavelength (nm)", ylab = "Signal", cex = 1.5)
}

}
