% Generated by tools/generate-man.R -- do not edit by hand
\name{latex}
\alias{latex}
\title{LaTeX Output Utilities}
\description{
Utilities for LaTeX output.
}
\usage{
latex(x)
}
\arguments{
  \item{x}{Object to format. }
}
\value{
LaTeX string. 
}
\seealso{

Other utilities: \code{\link{ascii}}, \code{\link{base}}, \code{\link{beep}}, \code{\link{bisection}}, \code{\link{cat.both}}, \code{\link{cat.stderr}}, \code{\link{chatter}}, \code{\link{check.names}}, \code{\link{debug.advice}}, \code{\link{debug.on}}, \code{\link{dict.lookup}}, \code{\link{done}}, \code{\link{enable.error.line.num}}, \code{\link{exit}}, \code{\link{grepcolname}}, \code{\link{heartbeat}}, \code{\link{index.of.variable}}, \code{\link{instrumentR}}, \code{\link{kable}}, \code{\link{latlon.distance}}, \code{\link{latlonx}}, \code{\link{mklatlonid}}, \code{\link{invlatlon}}, \code{\link{testlatlon}}, \code{\link{ls.objects}}, \code{\link{meminfo}}, \code{\link{mk.random.walk}}, \code{\link{mkRdictionary}}, \code{\link{msg}}, \code{\link{nearest}}, \code{\link{now}}, \code{\link{numobs.by}}, \code{\link{object.size.MB}}, \code{\link{osinfo}}, \code{\link{rdraw.perfect}}, \code{\link{require.variables}}, \code{\link{rexp.perfect}}, \code{\link{rnorm.perfect}}, \code{\link{runall}}, \code{\link{setupknitr}}, \code{\link{sink}}, \code{\link{sprint.data.frame}}, \code{\link{strcat}}, \code{\link{whatis}}, \code{\link{which.nearest}}
}
\examples{
\dontrun{
# Print a data frame as a LaTeX table (requires xtable)
df <- data.frame(x = 1:3, y = c(0.1, 0.2, 0.3))
iaw$latex(df)

# Print a matrix as a LaTeX table
m <- matrix(1:6, nrow = 2, dimnames = list(c("r1","r2"), c("a","b","c")))
iaw$latex(m)

# Regression summary table as LaTeX
fit <- lm(mpg ~ wt + hp, data = mtcars)
iaw$latex(summary(fit)$coefficients)

# Correlation matrix for inclusion in a paper
cormat <- cor(mtcars[, c("mpg", "wt", "hp", "disp")])
iaw$latex(round(cormat, 2))
}

}
