% Generated by tools/generate-man.R -- do not edit by hand
\name{latlon.distance}
\alias{latlon.distance}
\title{Latitude/Longitude Distance}
\description{
Calculates distance between lat/lon coordinates.
}
\usage{
latlon.distance(lat1, lon1, lat2, lon2)
}
\arguments{
  \item{lat1}{First latitude.}
  \item{lon1}{First longitude.}
  \item{lat2}{Second latitude.}
  \item{lon2}{Second longitude. }
}
\value{
Distance in kilometers. 
}
\seealso{

Other utilities: \code{\link{ascii}}, \code{\link{base}}, \code{\link{beep}}, \code{\link{bisection}}, \code{\link{cat.both}}, \code{\link{cat.stderr}}, \code{\link{chatter}}, \code{\link{check.names}}, \code{\link{debug.advice}}, \code{\link{debug.on}}, \code{\link{dict.lookup}}, \code{\link{done}}, \code{\link{enable.error.line.num}}, \code{\link{exit}}, \code{\link{grepcolname}}, \code{\link{heartbeat}}, \code{\link{index.of.variable}}, \code{\link{instrumentR}}, \code{\link{kable}}, \code{\link{latex}}, \code{\link{latlonx}}, \code{\link{mklatlonid}}, \code{\link{invlatlon}}, \code{\link{testlatlon}}, \code{\link{ls.objects}}, \code{\link{meminfo}}, \code{\link{mk.random.walk}}, \code{\link{mkRdictionary}}, \code{\link{msg}}, \code{\link{nearest}}, \code{\link{now}}, \code{\link{numobs.by}}, \code{\link{object.size.MB}}, \code{\link{osinfo}}, \code{\link{rdraw.perfect}}, \code{\link{require.variables}}, \code{\link{rexp.perfect}}, \code{\link{rnorm.perfect}}, \code{\link{runall}}, \code{\link{setupknitr}}, \code{\link{sink}}, \code{\link{sprint.data.frame}}, \code{\link{strcat}}, \code{\link{whatis}}, \code{\link{which.nearest}}
}
\examples{
\dontrun{
# Distance from New York to Los Angeles (~3940 km)
iaw$latlon.distance(40.7128, -74.0060, 34.0522, -118.2437)

# Distance from London to Paris (~341 km)
iaw$latlon.distance(51.5074, -0.1278, 48.8566, 2.3522)

# Vectorised: distances from one city to several others
iaw$latlon.distance(
  rep(40.7128, 3), rep(-74.0060, 3),
  c(34.0522, 41.8781, 29.7604),    # LA, Chicago, Houston
  c(-118.2437, -87.6298, -95.3698)
)
}
}
