% Generated by tools/generate-man.R -- do not edit by hand
\name{leadlagpanel}
\alias{leadlagpanel}
\title{Lead/Lag Panel Data}
\description{
Creates lead/lag variables for panel data.
}
\usage{
leadlagpanel(df, vars, nlags, panelid, timeid)
}
\arguments{
  \item{df}{Data frame.}
  \item{vars}{Variables to transform.}
  \item{nlags}{Number of lags (negative for leads).}
  \item{panelid}{Panel ID column name.}
  \item{timeid}{Time column name. }
}
\value{
Data frame with new columns. 
}
\seealso{

Other time-series: \code{\link{chgseries}}, \code{\link{compoundseries}}, \code{\link{cumulateseries}}, \code{\link{lagdataframe}}, \code{\link{lagseries}}, \code{\link{leadlagseries}}, \code{\link{leadseries}}, \code{\link{ma}}, \code{\link{pchgseries}}
}
\examples{
\dontrun{
panel <- data.frame(
  firm = c("A","A","A","B","B","B"),
  t    = c(1, 2, 3, 1, 2, 3),
  ret  = c(0.01, 0.02, 0.03, 0.10, 0.11, 0.12)
)

# One-period lag within each firm (adds column "ret.L1")
iaw$leadlagpanel(panel, vars = "ret", nlags = 1,
                 panelid = "firm", timeid = "t")

# One-period lead (negative nlags; adds column "ret.F1")
iaw$leadlagpanel(panel, vars = "ret", nlags = -1,
                 panelid = "firm", timeid = "t")

}
}
