% Generated by tools/generate-man.R -- do not edit by hand
\name{leadseries}
\alias{leadseries}
\title{Lead a Time Series}
\description{
Shifts series forward by numleads periods.
}
\usage{
leadseries(series, numleads = 1, panelid = NULL, timeid = NULL)
}
\arguments{
  \item{series}{A numeric vector.}
  \item{numleads}{Integer number of leads. Default 1.}
  \item{panelid}{Optional panel identifier vector.}
  \item{timeid}{Optional time identifier vector. }
}
\value{
Numeric vector of same length. 
}
\seealso{

Other time-series: \code{\link{chgseries}}, \code{\link{compoundseries}}, \code{\link{cumulateseries}}, \code{\link{lagdataframe}}, \code{\link{lagseries}}, \code{\link{leadlagpanel}}, \code{\link{leadlagseries}}, \code{\link{ma}}, \code{\link{pchgseries}}
}
\examples{
\dontrun{
x <- c(1, 2, 3, 4, 5)
iaw$leadseries(x)
}
}
