% Generated by tools/generate-man.R -- do not edit by hand
\name{long2wide}
\alias{long2wide}
\title{Reshape Long to Wide Format}
\description{
Converts long data frame to wide format.
}
\usage{
long2wide(d, id.time, id.firm, id.val)
}
\arguments{
  \item{d}{Data frame in long format.}
  \item{id.time}{Time variable (becomes columns).}
  \item{id.firm}{Entity variable (becomes rows).}
  \item{id.val}{Value variable. }
}
\value{
Data frame in wide format. 
}
\seealso{
tidyr::pivot_wider(d, names_from = id.time, values_from = id.val) 

Other data-reshaping: \code{\link{merge4seq}}, \code{\link{multimerge}}, \code{\link{wide2long}}
}
\examples{
\dontrun{
df <- data.frame(firm = c("A","A","B","B"), year = c(1,2,1,2), val = 1:4)
iaw$long2wide(df, "year", "firm", "val")
}
}
