% Generated by tools/generate-man.R -- do not edit by hand
\name{mc.replicate}
\alias{mc.replicate}
\title{Parallel Replicate}
\description{
Parallel version of replicate.
}
\usage{
mc.replicate(n, expr, simplify = TRUE)
}
\arguments{
  \item{n}{Number of replications.}
  \item{expr}{Expression to evaluate.}
  \item{simplify}{Simplify result. }
}
\value{
Vector or list of results. 
}
\seealso{

Other parallel: \code{\link{mc.by}}, \code{\link{mclapply}}, \code{\link{mcsapply}}, \code{\link{oc.by}}, \code{\link{rbind.mc.by}}, \code{\link{rbind.oc.by}}
}
\examples{
\dontrun{
# Estimate the sampling distribution of the mean via parallel simulation
means <- iaw$mc.replicate(1000, mean(rnorm(100)))
hist(means)

# Bootstrap a correlation coefficient
x <- rnorm(50); y <- x + rnorm(50)
boot_cors <- iaw$mc.replicate(500, {
  idx <- sample(50, replace = TRUE)
  cor(x[idx], y[idx])
})
quantile(boot_cors, c(0.025, 0.975))  # 95\% bootstrap CI

# Return a list (simplify = FALSE) when each replication yields a vector
sims <- iaw$mc.replicate(200, rnorm(5), simplify = FALSE)
length(sims)  # 200 elements, each of length 5

# Parallel Monte Carlo estimation of pi
pi_est <- iaw$mc.replicate(10000, {
  x <- runif(1); y <- runif(1)
  as.integer(x^2 + y^2 <= 1)
})
4 * mean(pi_est)  # approximately 3.14

# Simulate power of a t-test at a given effect size
rejections <- iaw$mc.replicate(1000, {
  x <- rnorm(30, mean = 0.3)
  t.test(x)$p.value < 0.05
})
mean(rejections)  # estimated power
}
}
