% Generated by tools/generate-man.R -- do not edit by hand
\name{mclapply}
\alias{mclapply}
\title{Enhanced Multicore Lapply}
\description{
Wrapper for parallel::mclapply with error handling.
}
\usage{
mclapply(X, FUN, ..., mc.cores = getOption("mc.cores", parallel::detectCores()), stop.on.error = TRUE, verbose = TRUE)
}
\arguments{
  \item{X}{List or vector.}
  \item{FUN}{Function to apply.}
  \item{...}{Arguments to FUN.}
  \item{mc.cores}{Number of cores.}
  \item{stop.on.error}{If TRUE, stop all workers on first error. Default TRUE.}
  \item{verbose}{Logical. Print error details on failure (default \code{TRUE}). }
}
\value{
List of results. 
}
\seealso{

Other parallel: \code{\link{mc.by}}, \code{\link{mc.replicate}}, \code{\link{mcsapply}}, \code{\link{oc.by}}, \code{\link{rbind.mc.by}}, \code{\link{rbind.oc.by}}
}
\examples{
\dontrun{
iaw$mclapply(1:10, function(x) x^2)
}
}
