% Generated by tools/generate-man.R -- do not edit by hand
\name{MixColor}
\alias{MixColor}
\alias{mixcolor}
\title{Mix Colors}
\description{
Mixes two colors.
}
\usage{
mixcolor(col1, col2, ratio = 0.5)
}
\arguments{
  \item{col1}{First color.}
  \item{col2}{Second color.}
  \item{ratio}{Mix ratio. }
}
\value{
Color value. 
}
\seealso{

Other plotting: \code{\link{as.PlotDate.yyyymmdd}}, \code{\link{plot.boxtext}}, \code{\link{colorblind}}, \code{\link{plot.contour}}, \code{\link{plot.ecdf.lineplot}}, \code{\link{plot.hist2d}}, \code{\link{hline}}, \code{\link{iaw.boxplot}}, \code{\link{plot.labelcurve}}, \code{\link{lambda2color}}, \code{\link{plot.loess.lineplot}}, \code{\link{plot.native.slope}}, \code{\link{plot.p.arrows}}, \code{\link{plot.pdf.embedfonts}}, \code{\link{pdf.end}}, \code{\link{plot.pdf.start.bera}}, \code{\link{pdf.start}}, \code{\link{plot.timearea}}, \code{\link{plot.worldmap}}, \code{\link{plot.points.toomany}}, \code{\link{plot.shade.xyy}}, \code{\link{plot.text.bg}}, \code{\link{plot.text.onbg}}, \code{\link{vline}}, \code{\link{plot.yearly.tickmarks}}, \code{\link{plot.ylab.left}}
}
\examples{
# 50/50 mix of red and blue
iaw$mixcolor("red", "blue")

# 25\% red, 75\% blue
iaw$mixcolor("red", "blue", ratio = 0.25)

# Pure endpoints: ratio 0 = col1, ratio 1 = col2
iaw$mixcolor("red", "blue", ratio = 0)  # "#FF0000" (pure red)
iaw$mixcolor("red", "blue", ratio = 1)  # "#0000FF" (pure blue)

# Mix black and white to produce a medium gray
iaw$mixcolor("black", "white", ratio = 0.5)  # "#808080"

# Create a custom heatmap color for a correlation value
r <- 0.73
col <- iaw$mixcolor("white", "darkred", ratio = abs(r))  # reddish for strong +r
nchar(col) > 0  # TRUE

\dontrun{
# Use mixed colors in a gradient legend
cols <- sapply(seq(0, 1, by = 0.1), function(r) iaw$mixcolor("red", "blue", r))
barplot(rep(1, length(cols)), col = cols, border = NA,
        main = "Red-to-blue gradient")

# Color bars by profit/loss magnitude using green-red mixing
returns <- c(0.05, -0.02, 0.10, -0.08, 0.03)
scaled <- (returns - min(returns)) / diff(range(returns))
bar_cols <- sapply(scaled, function(r) iaw$mixcolor("red", "forestgreen", r))
barplot(returns, col = bar_cols, names.arg = paste0("Q", 1:5),
        main = "Quarterly returns")
}

}
