% Generated by tools/generate-man.R -- do not edit by hand
\name{multimerge}
\alias{multimerge}
\title{Merge Multiple Data Frames}
\description{
Merges list of data frames.
}
\usage{
multimerge(dflist, by, ...)
}
\arguments{
  \item{dflist}{List of data frames.}
  \item{by}{Merge keys.}
  \item{...}{Additional merge arguments. }
}
\value{
Merged data frame. 
}
\seealso{

Other data-reshaping: \code{\link{long2wide}}, \code{\link{merge4seq}}, \code{\link{wide2long}}
}
\examples{
\dontrun{
d1 <- data.frame(id = 1:3, x = c(10, 20, 30))
d2 <- data.frame(id = 1:3, y = c(40, 50, 60))
d3 <- data.frame(id = 1:3, z = c(70, 80, 90))

# Merge three data frames on a common key
iaw$multimerge(list(d1, d2, d3), by = "id")

# Works with any number of frames; all.x = TRUE for left join behaviour
d4 <- data.frame(id = c(1, 2), w = c(100, 200))
iaw$multimerge(list(d1, d2, d4), by = "id", all.x = TRUE)

}
}
