% Generated by tools/generate-man.R -- do not edit by hand
\name{neweywest}
\alias{neweywest}
\title{Newey-West Standard Errors}
\description{
Wrapper for sandwich::NeweyWest.
}
\usage{
neweywest(model, lag = 0)
}
\arguments{
  \item{model}{Linear model object.}
  \item{lag}{Number of lags. }
}
\value{
Covariance matrix. 
}
\seealso{

Other regression: \code{\link{famamacbeth}}, \code{\link{olm}}, \code{\link{ols.nona}}, \code{\link{ooslm}}, \code{\link{print.olm}}, \code{\link{residuals}}
}
