% Generated by tools/generate-man.R -- do not edit by hand
\name{normalize}
\alias{normalize}
\title{Normalize a Vector}
\description{
Scales vector to [0, 1] range.
}
\usage{
normalize(x)
}
\arguments{
  \item{x}{Numeric vector. }
}
\value{
Numeric vector in [0, 1]. 
}
\seealso{

Other data-transformation: \code{\link{standardize}}, \code{\link{winsorize.level}}, \code{\link{winsorize.percentile}}
}
\examples{
\dontrun{
iaw$normalize(c(1, 2, 3, 4, 5))
}
}
