% Generated by tools/generate-man.R -- do not edit by hand
\name{object.size.MB}
\alias{object.size.MB}
\title{Object Size in MB}
\description{
Returns object size in megabytes.
}
\usage{
object.size.MB(x)
}
\arguments{
  \item{x}{Object to measure. }
}
\value{
Size in MB. 
}
\seealso{

Other utilities: \code{\link{ascii}}, \code{\link{base}}, \code{\link{beep}}, \code{\link{bisection}}, \code{\link{cat.both}}, \code{\link{cat.stderr}}, \code{\link{chatter}}, \code{\link{check.names}}, \code{\link{debug.advice}}, \code{\link{debug.on}}, \code{\link{dict.lookup}}, \code{\link{done}}, \code{\link{enable.error.line.num}}, \code{\link{exit}}, \code{\link{grepcolname}}, \code{\link{heartbeat}}, \code{\link{index.of.variable}}, \code{\link{instrumentR}}, \code{\link{kable}}, \code{\link{latex}}, \code{\link{latlon.distance}}, \code{\link{latlonx}}, \code{\link{mklatlonid}}, \code{\link{invlatlon}}, \code{\link{testlatlon}}, \code{\link{ls.objects}}, \code{\link{meminfo}}, \code{\link{mk.random.walk}}, \code{\link{mkRdictionary}}, \code{\link{msg}}, \code{\link{nearest}}, \code{\link{now}}, \code{\link{numobs.by}}, \code{\link{osinfo}}, \code{\link{rdraw.perfect}}, \code{\link{require.variables}}, \code{\link{rexp.perfect}}, \code{\link{rnorm.perfect}}, \code{\link{runall}}, \code{\link{setupknitr}}, \code{\link{sink}}, \code{\link{sprint.data.frame}}, \code{\link{strcat}}, \code{\link{whatis}}, \code{\link{which.nearest}}
}
\examples{
\dontrun{
iaw$object.size.MB(1:1000000)          # ~3.81 MB

# Small object
iaw$object.size.MB(1:10)               # ~0 MB

# Compare storage of different types
iaw$object.size.MB(double(1e6))        # 8 bytes each -> ~7.63 MB
iaw$object.size.MB(integer(1e6))       # 4 bytes each -> ~3.81 MB

# Check size of a data frame
d <- data.frame(x = rnorm(1e5), y = rnorm(1e5), z = sample(letters, 1e5, replace = TRUE))
iaw$object.size.MB(d)

# Character vectors can be surprisingly large
big_str <- replicate(1e4, paste(sample(letters, 50, TRUE), collapse = ""))
iaw$object.size.MB(big_str)

# Logical vectors are compact
iaw$object.size.MB(rep(TRUE, 1e6))   # ~3.81 MB (same as integer)

# Matrix vs data frame overhead
m <- matrix(0, nrow = 1000, ncol = 100)
iaw$object.size.MB(m)                      # raw storage
iaw$object.size.MB(as.data.frame(m))       # slightly larger due to column list
}
}
