% Generated by tools/generate-man.R -- do not edit by hand
\name{olm}
\alias{olm}
\title{OLM}
\description{
Run an ordinary-least-squares model with enhanced output.
}
\usage{
olm(..., newey.west = 0, stdcoefs = TRUE, include.anova = TRUE, keep.pval = FALSE, truncate.T = TRUE, digits = 4)
}
\arguments{
  \item{...}{Arguments passed to lm().}
  \item{newey.west}{Lag for Newey-West standard errors. Default 0.}
  \item{stdcoefs}{Include standardized coefficients. Default TRUE.}
  \item{include.anova}{Include ANOVA sum of squares decomposition. Default TRUE.}
  \item{keep.pval}{Keep p-values in output. Default FALSE.}
  \item{truncate.T}{Truncate t-statistics to +/-99.99. Default TRUE.}
  \item{digits}{Rounding digits. Default 4. }
}
\value{
An object of class "olm" inheriting from "summary.lm". 
}
\seealso{
print.olm

Other regression: \code{\link{famamacbeth}}, \code{\link{neweywest}}, \code{\link{ols.nona}}, \code{\link{ooslm}}, \code{\link{print.olm}}, \code{\link{residuals}}
}
