% Generated by tools/generate-man.R -- do not edit by hand
\name{pchgseries}
\alias{pchgseries}
\title{Calculate Percent Changes in a Time Series}
\description{
Computes percent change: x[t]/x[t-n] - 1.
}
\usage{
pchgseries(seriesin, numlags = 1, panelid = NULL, timeid = NULL)
}
\arguments{
  \item{seriesin}{Numeric vector.}
  \item{numlags}{Number of periods. Default 1.}
  \item{panelid}{Optional panel identifier.}
  \item{timeid}{Optional time identifier. }
}
\value{
Numeric vector of same length. 
}
\seealso{

Other time-series: \code{\link{chgseries}}, \code{\link{compoundseries}}, \code{\link{cumulateseries}}, \code{\link{lagdataframe}}, \code{\link{lagseries}}, \code{\link{leadlagpanel}}, \code{\link{leadlagseries}}, \code{\link{leadseries}}, \code{\link{ma}}
}
\examples{
\dontrun{
prices <- c(100, 105, 103, 110)
iaw$pchgseries(prices)
}
}
