% Generated by tools/generate-man.R -- do not edit by hand
\name{\%inrange\%}
\alias{\%inrange\%}
\title{Check if Values are Within a Range}
\description{
Tests whether values fall within a specified inclusive range.
}
\usage{
e1 \%inrange\% e2
}
\arguments{
  \item{x}{A numeric value or vector to test.}
  \item{range_vector}{A numeric vector of length 2: c(min, max). }
}
\value{
A logical vector. 
}
\seealso{

Other operators: \code{\link{\%and\%}}, \code{\link{\%or\%}}
}
\examples{
\dontrun{
5 \%inrange\% c(1, 10)
c(-5, 0, 5, 10, 15) \%inrange\% c(0, 10)
}
}
