% Generated by tools/generate-man.R -- do not edit by hand
\name{plot.native.slope}
\alias{plot.native.slope}
\alias{native.slope}
\title{Calculate Native Slope Angle}
\description{
Calculates rotation angle for text along slope.
}
\usage{
native.slope(slope)
}
\arguments{
  \item{slope}{Slope value. }
}
\value{
Angle in degrees. 
}
\seealso{

Other plotting: \code{\link{as.PlotDate.yyyymmdd}}, \code{\link{plot.boxtext}}, \code{\link{colorblind}}, \code{\link{plot.contour}}, \code{\link{plot.ecdf.lineplot}}, \code{\link{plot.hist2d}}, \code{\link{hline}}, \code{\link{iaw.boxplot}}, \code{\link{plot.labelcurve}}, \code{\link{lambda2color}}, \code{\link{plot.loess.lineplot}}, \code{\link{MixColor}}, \code{\link{plot.p.arrows}}, \code{\link{plot.pdf.embedfonts}}, \code{\link{pdf.end}}, \code{\link{pdf.start}}, \code{\link{plot.pdf.start.bera}}, \code{\link{plot.timearea}}, \code{\link{plot.worldmap}}, \code{\link{plot.points.toomany}}, \code{\link{plot.shade.xyy}}, \code{\link{plot.text.bg}}, \code{\link{plot.text.onbg}}, \code{\link{vline}}, \code{\link{plot.yearly.tickmarks}}, \code{\link{plot.ylab.left}}
}
\examples{
\dontrun{
# Get the visual rotation angle for a slope of 1 (45 degrees on a square plot)
plot(1:10, 1:10)
angle <- iaw$native.slope(1)
text(5, 5, "slope=1", srt = angle)

# Label a regression line at the correct visual angle
x <- 1:20
y <- 0.5 * x + rnorm(20)
fit <- lm(y ~ x)
plot(x, y)
abline(fit)
srt <- iaw$native.slope(coef(fit)[["x"]])
text(10, predict(fit, data.frame(x = 10)), "fit", srt = srt, pos = 3)
}

}
