% Generated by tools/generate-man.R -- do not edit by hand
\name{plot.pdf.embedfonts}
\alias{plot.pdf.embedfonts}
\alias{pdf.embedfonts}
\title{Embed Fonts in PDF}
\description{
Embeds fonts in a PDF file.
}
\usage{
pdf.embedfonts(filename)
}
\arguments{
  \item{filename}{PDF filename. }
}
\value{
Invisible NULL. 
}
\seealso{

Other plotting: \code{\link{as.PlotDate.yyyymmdd}}, \code{\link{plot.boxtext}}, \code{\link{colorblind}}, \code{\link{plot.contour}}, \code{\link{plot.ecdf.lineplot}}, \code{\link{plot.hist2d}}, \code{\link{hline}}, \code{\link{iaw.boxplot}}, \code{\link{plot.labelcurve}}, \code{\link{lambda2color}}, \code{\link{plot.loess.lineplot}}, \code{\link{MixColor}}, \code{\link{plot.native.slope}}, \code{\link{plot.p.arrows}}, \code{\link{pdf.end}}, \code{\link{pdf.start}}, \code{\link{plot.pdf.start.bera}}, \code{\link{plot.timearea}}, \code{\link{plot.worldmap}}, \code{\link{plot.points.toomany}}, \code{\link{plot.shade.xyy}}, \code{\link{plot.text.bg}}, \code{\link{plot.text.onbg}}, \code{\link{vline}}, \code{\link{plot.yearly.tickmarks}}, \code{\link{plot.ylab.left}}
}
\examples{
\dontrun{
# Create a PDF then embed fonts for portability
iaw$pdf.start("myplot.pdf")
plot(1:10, main = "Test plot")
iaw$pdf.end()
iaw$pdf.embedfonts("myplot.pdf")

# Embed fonts in an existing PDF file
iaw$pdf.embedfonts("results/figure1.pdf")
}

}
