% Generated by tools/generate-man.R -- do not edit by hand
\name{plot.yearly.tickmarks}
\alias{plot.yearly.tickmarks}
\title{Add Yearly Tickmarks}
\description{
Adds yearly tick marks to time axis.
}
\usage{
plot.yearly.tickmarks(years, ...)
}
\arguments{
  \item{years}{Years to mark.}
  \item{...}{Axis arguments. }
}
\value{
Invisible NULL. 
}
\seealso{

Other plotting: \code{\link{as.PlotDate.yyyymmdd}}, \code{\link{plot.boxtext}}, \code{\link{colorblind}}, \code{\link{plot.contour}}, \code{\link{plot.ecdf.lineplot}}, \code{\link{plot.hist2d}}, \code{\link{hline}}, \code{\link{iaw.boxplot}}, \code{\link{plot.labelcurve}}, \code{\link{lambda2color}}, \code{\link{plot.loess.lineplot}}, \code{\link{MixColor}}, \code{\link{plot.native.slope}}, \code{\link{plot.p.arrows}}, \code{\link{plot.pdf.embedfonts}}, \code{\link{pdf.end}}, \code{\link{pdf.start}}, \code{\link{plot.pdf.start.bera}}, \code{\link{plot.timearea}}, \code{\link{plot.worldmap}}, \code{\link{plot.points.toomany}}, \code{\link{plot.shade.xyy}}, \code{\link{plot.text.bg}}, \code{\link{plot.text.onbg}}, \code{\link{vline}}, \code{\link{plot.ylab.left}}
}
\examples{
\dontrun{
# Draw a time series and add yearly tick marks
plot(2000:2010, cumsum(rnorm(11)), type = "l", xaxt = "n")
iaw$plot.yearly.tickmarks(2000:2010)

# Add tick marks for every other year with smaller text
plot(2000:2020, cumsum(rnorm(21)), type = "l", xaxt = "n")
iaw$plot.yearly.tickmarks(seq(2000, 2020, by = 2), cex.axis = 0.8)
}

}
