% Generated by tools/generate-man.R -- do not edit by hand
\name{residuals}
\alias{residuals}
\title{Extract Residuals}
\description{
Extracts residuals from model.
}
\usage{
residuals(model, ...)
}
\arguments{
  \item{model}{Regression model (from \code{lm} or similar).}
  \item{...}{Additional arguments passed to \code{stats::residuals}. }
}
\value{
Residuals vector (with NAs preserved if \code{na.exclude} was used). 
}
\seealso{

Other regression: \code{\link{famamacbeth}}, \code{\link{neweywest}}, \code{\link{olm}}, \code{\link{ols.nona}}, \code{\link{ooslm}}, \code{\link{print.olm}}
}
