% Generated by tools/generate-man.R -- do not edit by hand
\name{setupknitr}
\alias{setupknitr}
\title{Setup Knitr Options}
\description{
Sets default knitr chunk options.
}
\usage{
setupknitr()
}
\value{
Invisible NULL. 
}
\seealso{

Other utilities: \code{\link{ascii}}, \code{\link{base}}, \code{\link{beep}}, \code{\link{bisection}}, \code{\link{cat.both}}, \code{\link{cat.stderr}}, \code{\link{chatter}}, \code{\link{check.names}}, \code{\link{debug.advice}}, \code{\link{debug.on}}, \code{\link{dict.lookup}}, \code{\link{done}}, \code{\link{enable.error.line.num}}, \code{\link{exit}}, \code{\link{grepcolname}}, \code{\link{heartbeat}}, \code{\link{index.of.variable}}, \code{\link{instrumentR}}, \code{\link{kable}}, \code{\link{latex}}, \code{\link{latlon.distance}}, \code{\link{latlonx}}, \code{\link{mklatlonid}}, \code{\link{invlatlon}}, \code{\link{testlatlon}}, \code{\link{ls.objects}}, \code{\link{meminfo}}, \code{\link{mk.random.walk}}, \code{\link{mkRdictionary}}, \code{\link{msg}}, \code{\link{nearest}}, \code{\link{now}}, \code{\link{numobs.by}}, \code{\link{object.size.MB}}, \code{\link{osinfo}}, \code{\link{rdraw.perfect}}, \code{\link{require.variables}}, \code{\link{rexp.perfect}}, \code{\link{rnorm.perfect}}, \code{\link{runall}}, \code{\link{sink}}, \code{\link{sprint.data.frame}}, \code{\link{strcat}}, \code{\link{whatis}}, \code{\link{which.nearest}}
}
\examples{
\dontrun{
# Call at the top of an R Markdown document to apply standard chunk defaults
iaw$setupknitr()

# After calling, individual chunks can still override individual options
# ```{r, fig.width=10}
# ... wider plot ...
# ```

# Typical Rmd setup chunk:
# ```{r setup, include=FALSE}
# library(iaw)
# iaw$setupknitr()
# ```

# Sets echo=TRUE, message=FALSE, warning=FALSE, fig.width=7, fig.height=5
# These are sensible defaults for financial reports and data analysis
}

}
