% Generated by tools/generate-man.R -- do not edit by hand
\name{standardize}
\alias{standardize}
\title{Standardize a Vector}
\description{
Centers and scales to mean=0, sd=1.
}
\usage{
standardize(x)
}
\arguments{
  \item{x}{Numeric vector. }
}
\value{
Numeric vector. 
}
\seealso{

Other data-transformation: \code{\link{normalize}}, \code{\link{winsorize.level}}, \code{\link{winsorize.percentile}}
}
\examples{
\dontrun{
iaw$standardize(c(1, 2, 3, 4, 5))
}
}
