% Generated by tools/generate-man.R -- do not edit by hand
\name{summary}
\alias{summary}
\title{Descriptive Statistics Summary}
\description{
Computes comprehensive descriptive statistics.
}
\usage{
summary(df, verbose = "X", digits = 4)
}
\arguments{
  \item{df}{Data frame, matrix, or vector.}
  \item{verbose}{Statistics profile: "p", "x", "X", "a", "sr", "sr252".}
  \item{digits}{Decimal places. Default 4. }
}
\value{
Numeric matrix of statistics. 
}
\seealso{

Other statistics: \code{\link{autocorrel}}, \code{\link{autopcorrel}}, \code{\link{cmpsum}}, \code{\link{colN}}, \code{\link{colSds}}, \code{\link{covp}}, \code{\link{dependent.rank}}, \code{\link{print.cor}}, \code{\link{rank}}, \code{\link{rowSds}}, \code{\link{sdp}}, \code{\link{sse}}, \code{\link{sst}}, \code{\link{tabularsummary}}, \code{\link{varp}}
}
\examples{
\dontrun{
df <- data.frame(x = rnorm(100), y = rnorm(100))
iaw$summary(df, "p")
}
}
