% Generated by tools/generate-man.R -- do not edit by hand
\name{tabularsummary}
\alias{tabularsummary}
\title{Create Tabular Summary}
\description{
Creates frequency table with statistics.
}
\usage{
tabularsummary(x, sort = TRUE)
}
\arguments{
  \item{x}{Factor or character vector.}
  \item{sort}{Sort by frequency. }
}
\value{
Data frame with counts and percentages. 
}
\seealso{

Other statistics: \code{\link{autocorrel}}, \code{\link{autopcorrel}}, \code{\link{cmpsum}}, \code{\link{colN}}, \code{\link{colSds}}, \code{\link{covp}}, \code{\link{dependent.rank}}, \code{\link{print.cor}}, \code{\link{rank}}, \code{\link{rowSds}}, \code{\link{sdp}}, \code{\link{sse}}, \code{\link{sst}}, \code{\link{summary}}, \code{\link{varp}}
}
\examples{
\dontrun{
iaw$tabularsummary(c("a", "b", "a", "c", "a"))
}
}
