% Generated by tools/generate-man.R -- do not edit by hand
\name{tmdiffsec}
\alias{tmdiffsec}
\title{Time Difference in Seconds}
\description{
Calculates time difference in seconds.
}
\usage{
tmdiffsec(t1, t2)
}
\arguments{
  \item{t1}{First time.}
  \item{t2}{Second time. }
}
\value{
Numeric seconds. 
}
\seealso{

Other datetime: \code{\link{diffdays}}, \code{\link{epoch2nyc}}, \code{\link{epoch2POSIXct}}, \code{\link{epoch2yyyymmdd}}, \code{\link{yyyymmdd.toggle}}, \code{\link{yyyymmdd}}
}
\examples{
\dontrun{
t1 <- as.POSIXct("2024-01-01 10:00:00")
t2 <- as.POSIXct("2024-01-01 10:05:30")

# Elapsed seconds between two timestamps
iaw$tmdiffsec(t1, t2)   # 330

# Measure wall-clock time for a computation
start <- Sys.time()
Sys.sleep(0.1)
iaw$tmdiffsec(start, Sys.time())   # approximately 0.1

# Difference across midnight
t_before <- as.POSIXct("2024-06-15 23:59:50")
t_after  <- as.POSIXct("2024-06-16 00:00:10")
iaw$tmdiffsec(t_before, t_after)   # 20

# Negative result when t2 is earlier than t1
iaw$tmdiffsec(t_after, t_before)   # -20

# Convert to minutes by dividing
t1 <- as.POSIXct("2024-01-01 09:00:00")
t2 <- as.POSIXct("2024-01-01 10:30:00")
iaw$tmdiffsec(t1, t2) / 60   # 90 minutes

}
}
