% Generated by tools/generate-man.R -- do not edit by hand
\name{touch}
\alias{touch}
\title{Touch File}
\description{
Creates empty file or updates timestamp.
}
\usage{
touch(filename)
}
\arguments{
  \item{filename}{File path. }
}
\value{
Invisible TRUE. 
}
\seealso{

Other io: \code{\link{lock}}, \code{\link{read.csv}}, \code{\link{.Riolog}}, \code{\link{write.csv}}
}
