% Generated by tools/generate-man.R -- do not edit by hand
\name{varp}
\alias{varp}
\title{Population Variance}
\description{
Calculates population variance (n denominator).
}
\usage{
varp(x, na.rm = TRUE)
}
\arguments{
  \item{x}{Numeric vector.}
  \item{na.rm}{Remove NA values. }
}
\value{
Variance value. 
}
\seealso{

Other statistics: \code{\link{autocorrel}}, \code{\link{autopcorrel}}, \code{\link{cmpsum}}, \code{\link{colN}}, \code{\link{colSds}}, \code{\link{covp}}, \code{\link{dependent.rank}}, \code{\link{print.cor}}, \code{\link{rank}}, \code{\link{rowSds}}, \code{\link{sdp}}, \code{\link{sse}}, \code{\link{sst}}, \code{\link{summary}}, \code{\link{tabularsummary}}
}
\examples{
\dontrun{
iaw$varp(1:10)
}
}
