% Generated by tools/generate-man.R -- do not edit by hand
\name{whatis}
\alias{whatis}
\title{Describe Object Type and Structure}
\description{
Returns description of object type, class, and dimensions.
}
\usage{
whatis(x)
}
\arguments{
  \item{x}{Any R object. }
}
\value{
Character string description. 
}
\seealso{

Other utilities: \code{\link{ascii}}, \code{\link{base}}, \code{\link{beep}}, \code{\link{bisection}}, \code{\link{cat.both}}, \code{\link{cat.stderr}}, \code{\link{chatter}}, \code{\link{check.names}}, \code{\link{debug.advice}}, \code{\link{debug.on}}, \code{\link{dict.lookup}}, \code{\link{done}}, \code{\link{enable.error.line.num}}, \code{\link{exit}}, \code{\link{grepcolname}}, \code{\link{heartbeat}}, \code{\link{index.of.variable}}, \code{\link{instrumentR}}, \code{\link{kable}}, \code{\link{latex}}, \code{\link{latlon.distance}}, \code{\link{latlonx}}, \code{\link{mklatlonid}}, \code{\link{invlatlon}}, \code{\link{testlatlon}}, \code{\link{ls.objects}}, \code{\link{meminfo}}, \code{\link{mk.random.walk}}, \code{\link{mkRdictionary}}, \code{\link{msg}}, \code{\link{nearest}}, \code{\link{now}}, \code{\link{numobs.by}}, \code{\link{object.size.MB}}, \code{\link{osinfo}}, \code{\link{rdraw.perfect}}, \code{\link{require.variables}}, \code{\link{rexp.perfect}}, \code{\link{rnorm.perfect}}, \code{\link{runall}}, \code{\link{setupknitr}}, \code{\link{sink}}, \code{\link{sprint.data.frame}}, \code{\link{strcat}}, \code{\link{which.nearest}}
}
\examples{
\dontrun{
iaw$whatis(1:10)                        # "integer (integer) length=10"
iaw$whatis(data.frame(a = 1))           # "data.frame (list) [1x1]"

# Matrix shows dimensions
iaw$whatis(matrix(1:6, nrow = 2))       # "matrix (integer) [2x3]"

# List and logical
iaw$whatis(list(a = 1, b = "x"))        # "list (list) length=2"
iaw$whatis(c(TRUE, FALSE, TRUE))        # "logical (logical) length=3"

# Useful for quick inspection inside scripts
x <- rnorm(50)
message(iaw$whatis(x))                  # "numeric (double) length=50"
}
}
