% Generated by tools/generate-man.R -- do not edit by hand
\name{winsorize.percentile}
\alias{winsorize.percentile}
\title{Winsorize at Percentile Thresholds}
\description{
Clips values at specified percentiles.
}
\usage{
winsorize.percentile(x, xperc = c(0.01, 0.99), verbose = FALSE, name = "")
}
\arguments{
  \item{x}{Numeric vector.}
  \item{xperc}{vector of Lower percentile (0-1). Default 0.01, upper percentile}
  \item{verbose}{Print summary.}
  \item{name}{Variable name for output. }
}
\value{
Numeric vector. 
}
\seealso{

Other data-transformation: \code{\link{normalize}}, \code{\link{standardize}}, \code{\link{winsorize.level}}
}
\examples{
\dontrun{
x <- c(rnorm(98), -100, 100)
iaw$winsorize.percentile(x)
}
}
