% Generated by tools/generate-man.R -- do not edit by hand
\name{yyyymmdd}
\alias{yyyymmdd}
\title{Convert Date-Like Input to Various Formats}
\description{
Converts YYYYMMDD integers, Gregorian day numbers, or POSIXct objects
to the requested output format.
}
\usage{
yyyymmdd(anything, output=c("posix", "gregorian", "yyyymmdd", "weeknum", "weekday"))
}
\arguments{
  \item{anything}{Numeric vector (YYYYMMDD or Gregorian) or POSIXct.}
  \item{output}{Output format: one of \code{"posix"}, \code{"gregorian"}, \code{"yyyymmdd"}, \code{"weeknum"}, or \code{"weekday"}. }
}
\value{
Converted date in the requested format. 
}
\seealso{
[iaw$yyyymmdd.toggle()], [iaw$as.PlotDate.yyyymmdd()] 

Other datetime: \code{\link{diffdays}}, \code{\link{epoch2nyc}}, \code{\link{epoch2POSIXct}}, \code{\link{epoch2yyyymmdd}}, \code{\link{tmdiffsec}}, \code{\link{yyyymmdd.toggle}}
}
\examples{
\dontrun{
iaw$yyyymmdd(20240115, output = "weekday")
}
}
