% Generated by tools/generate-man.R -- do not edit by hand
\name{yyyymmdd.toggle}
\alias{yyyymmdd.toggle}
\title{Toggle Between YYYYMMDD Integer and Gregorian Day Number}
\description{
Converts YYYYMMDD integers to Gregorian day numbers (days since 1970-01-01)
and vice versa. Autodetects direction by value range.
}
\usage{
yyyymmdd.toggle(yyyymmdd_or_int)
}
\arguments{
  \item{yyyymmdd_or_int}{Numeric vector of YYYYMMDD integers (>= 19000000) or Gregorian day numbers (-300000 to 50000). }
}
\value{
Numeric vector converted to the other format. 
}
\seealso{

Other datetime: \code{\link{diffdays}}, \code{\link{epoch2nyc}}, \code{\link{epoch2POSIXct}}, \code{\link{epoch2yyyymmdd}}, \code{\link{tmdiffsec}}, \code{\link{yyyymmdd}}
}
\examples{
\dontrun{
iaw$yyyymmdd.toggle(20210115)
iaw$yyyymmdd.toggle(18642)  # back to YYYYMMDD
}
}
