% Generated by tools/generate-man.R -- do not edit by hand
\name{autocorrel}
\alias{autocorrel}
\title{Cross-Correlation at Various Lags}
\description{
Calculates cross-correlations at different leads/lags.
**WARNING**: Be careful that you understand what the lead and what the lag is
}
\usage{
autocorrel(series.x, series.y, leadlags = 5)
}
\arguments{
  \item{series.x}{First numeric vector.}
  \item{series.y}{Second numeric vector.}
  \item{leadlags}{Number of leads/lags (default 5). }
}
\value{
Named numeric vector of correlations. 
}
\seealso{

Other statistics: \code{\link{autopcorrel}}, \code{\link{cmpsum}}, \code{\link{colN}}, \code{\link{colSds}}, \code{\link{covp}}, \code{\link{dependent.rank}}, \code{\link{print.cor}}, \code{\link{rank}}, \code{\link{rowSds}}, \code{\link{sdp}}, \code{\link{sse}}, \code{\link{sst}}, \code{\link{summary}}, \code{\link{tabularsummary}}, \code{\link{varp}}
}
\examples{
\dontrun{
x <- sin(1:100)
y <- sin(2:101)
iaw$autocorrel(x, y, leadlags = 3)
}
}
