% Generated by tools/generate-man.R -- do not edit by hand
\name{autopcorrel}
\alias{autopcorrel}
\title{Partial Cross-Correlation at Various Lags}
\description{
Calculates partial cross-correlations using regression.
}
\usage{
autopcorrel(series.x, series.y, around = 5)
}
\arguments{
  \item{series.x}{First numeric vector.}
  \item{series.y}{Second numeric vector.}
  \item{around}{Number of leads/lags (default 5). }
}
\value{
Named numeric vector of partial correlations. 
}
\seealso{

Other statistics: \code{\link{autocorrel}}, \code{\link{cmpsum}}, \code{\link{colN}}, \code{\link{colSds}}, \code{\link{covp}}, \code{\link{dependent.rank}}, \code{\link{print.cor}}, \code{\link{rank}}, \code{\link{rowSds}}, \code{\link{sdp}}, \code{\link{sse}}, \code{\link{sst}}, \code{\link{summary}}, \code{\link{tabularsummary}}, \code{\link{varp}}
}
\examples{
\dontrun{
x <- sin(1:50) + rnorm(50, sd = 0.1)
y <- sin(2:51) + rnorm(50, sd = 0.1)
iaw$autopcorrel(x, y, around = 3)
}
}
