% Generated by tools/generate-man.R -- do not edit by hand
\name{bisection}
\alias{bisection}
\title{Bisection Root Finding (Vectorized)}
\description{
Finds roots using bisection method. Supports vectorized inputs.
}
\usage{
bisection(f, lower, upper, ..., numiter = 100, tolerance = .Machine$double.eps^0.25, verbose = 0)
}
\arguments{
  \item{f}{Function to find root of.}
  \item{lower}{Numeric vector of lower bounds.}
  \item{upper}{Numeric vector of upper bounds.}
  \item{...}{Additional arguments passed to f.}
  \item{numiter}{Maximum iterations. Default 100.}
  \item{tolerance}{Convergence tolerance. Default .Machine$double.eps^0.25.}
  \item{verbose}{Print progress. Default 0. }
}
\value{
List with mid, fmid, lower, upper, flower, fupper, n. 
}
\seealso{

Other utilities: \code{\link{ascii}}, \code{\link{base}}, \code{\link{beep}}, \code{\link{cat.both}}, \code{\link{cat.stderr}}, \code{\link{chatter}}, \code{\link{check.names}}, \code{\link{debug.advice}}, \code{\link{debug.on}}, \code{\link{dict.lookup}}, \code{\link{done}}, \code{\link{enable.error.line.num}}, \code{\link{exit}}, \code{\link{grepcolname}}, \code{\link{heartbeat}}, \code{\link{index.of.variable}}, \code{\link{instrumentR}}, \code{\link{kable}}, \code{\link{latex}}, \code{\link{latlon.distance}}, \code{\link{latlonx}}, \code{\link{mklatlonid}}, \code{\link{invlatlon}}, \code{\link{testlatlon}}, \code{\link{ls.objects}}, \code{\link{meminfo}}, \code{\link{mk.random.walk}}, \code{\link{mkRdictionary}}, \code{\link{msg}}, \code{\link{nearest}}, \code{\link{now}}, \code{\link{numobs.by}}, \code{\link{object.size.MB}}, \code{\link{osinfo}}, \code{\link{rdraw.perfect}}, \code{\link{require.variables}}, \code{\link{rexp.perfect}}, \code{\link{rnorm.perfect}}, \code{\link{runall}}, \code{\link{setupknitr}}, \code{\link{sink}}, \code{\link{sprint.data.frame}}, \code{\link{strcat}}, \code{\link{whatis}}, \code{\link{which.nearest}}
}
\examples{
\dontrun{
iaw$bisection(function(x) x^2 - 2, 0, 2)
}
}
