% Generated by tools/generate-man.R -- do not edit by hand
\name{chgseries}
\alias{chgseries}
\alias{dchgseries}
\title{Calculate Changes in a Time Series}
\description{
Computes first difference: x[t] - x[t-n].
}
\usage{
chgseries(seriesin, numlags = 1, panelid = NULL, timeid = NULL)
}
\arguments{
  \item{seriesin}{Numeric vector.}
  \item{numlags}{Number of periods. Default 1.}
  \item{panelid}{Optional panel identifier.}
  \item{timeid}{Optional time identifier. }
}
\value{
Numeric vector of same length. 
}
\seealso{

Other time-series: \code{\link{compoundseries}}, \code{\link{cumulateseries}}, \code{\link{lagdataframe}}, \code{\link{lagseries}}, \code{\link{leadlagpanel}}, \code{\link{leadlagseries}}, \code{\link{leadseries}}, \code{\link{ma}}, \code{\link{pchgseries}}
}
\examples{
\dontrun{
x <- c(100, 102, 105, 103, 108)
iaw$chgseries(x)
}
}
