% Generated by tools/generate-man.R -- do not edit by hand
\name{colN}
\alias{colN}
\title{Column Count Non-NA}
\description{
Counts non-NA values in each column.
}
\usage{
colN(x)
}
\arguments{
  \item{x}{Matrix or data frame. }
}
\value{
Vector of counts. 
}
\seealso{

Other statistics: \code{\link{autocorrel}}, \code{\link{autopcorrel}}, \code{\link{cmpsum}}, \code{\link{colSds}}, \code{\link{covp}}, \code{\link{dependent.rank}}, \code{\link{print.cor}}, \code{\link{rank}}, \code{\link{rowSds}}, \code{\link{sdp}}, \code{\link{sse}}, \code{\link{sst}}, \code{\link{summary}}, \code{\link{tabularsummary}}, \code{\link{varp}}
}
\examples{
\dontrun{
df <- data.frame(a = c(1, NA, 3), b = c(1, 2, 3))
iaw$colN(df)
}
}
