% Generated by tools/generate-man.R -- do not edit by hand
\name{colSds}
\alias{colSds}
\title{Column Standard Deviations}
\description{
Calculates SD for each column.
}
\usage{
colSds(x, na.rm)
}
\arguments{
  \item{x}{Matrix or data frame.}
  \item{na.rm}{Remove NA values. }
}
\value{
Vector of standard deviations. 
}
\seealso{

Other statistics: \code{\link{autocorrel}}, \code{\link{autopcorrel}}, \code{\link{cmpsum}}, \code{\link{colN}}, \code{\link{covp}}, \code{\link{dependent.rank}}, \code{\link{print.cor}}, \code{\link{rank}}, \code{\link{rowSds}}, \code{\link{sdp}}, \code{\link{sse}}, \code{\link{sst}}, \code{\link{summary}}, \code{\link{tabularsummary}}, \code{\link{varp}}
}
\examples{
\dontrun{
iaw$colSds(matrix(rnorm(20), ncol = 4))
}
}
