% Generated by tools/generate-man.R -- do not edit by hand
\name{covp}
\alias{covp}
\title{Population Covariance}
\description{
Calculates population covariance (n denominator).
}
\usage{
covp(x, y, na.rm = TRUE)
}
\arguments{
  \item{x}{First vector.}
  \item{y}{Second vector.}
  \item{na.rm}{Remove NA values. }
}
\value{
Covariance value. 
}
\seealso{

Other statistics: \code{\link{autocorrel}}, \code{\link{autopcorrel}}, \code{\link{cmpsum}}, \code{\link{colN}}, \code{\link{colSds}}, \code{\link{dependent.rank}}, \code{\link{print.cor}}, \code{\link{rank}}, \code{\link{rowSds}}, \code{\link{sdp}}, \code{\link{sse}}, \code{\link{sst}}, \code{\link{summary}}, \code{\link{tabularsummary}}, \code{\link{varp}}
}
\examples{
\dontrun{
iaw$covp(1:10, 11:20)
}
}
