% Generated by tools/generate-man.R -- do not edit by hand
\name{covp}
\alias{covp}
\title{Population Covariance}
\description{
Calculates population covariance (n denominator).
}
\usage{
covp(x, y, na.rm = TRUE)
}
\arguments{
  \item{x}{First vector.}
  \item{y}{Second vector.}
  \item{na.rm}{Remove NA values. }
}
\value{
Covariance value. 
}
\seealso{

Other statistics: \code{\link{autocorrel}}, \code{\link{autopcorrel}}, \code{\link{cmpsum}}, \code{\link{colN}}, \code{\link{colSds}}, \code{\link{dependent.rank}}, \code{\link{print.cor}}, \code{\link{rank}}, \code{\link{rowSds}}, \code{\link{sdp}}, \code{\link{sse}}, \code{\link{sst}}, \code{\link{summary}}, \code{\link{tabularsummary}}, \code{\link{varp}}
}
\examples{
\dontrun{
# Population covariance (n denominator) vs sample covariance (n-1)
x <- 1:10
iaw$covp(x, x)   # population variance of x with itself
cov(x, x)        # sample covariance (slightly larger)

# Two related series
iaw$covp(c(1, 2, 3, 4, 5), c(2, 4, 5, 4, 5))

# NA values are removed pairwise by default
iaw$covp(c(1, NA, 3, 4), c(2, 3, NA, 5))

# All columns of a matrix against a reference vector
m <- matrix(c(1:5, 5:1, c(1,3,5,3,1)), ncol = 3)
apply(m, 2, function(col) iaw$covp(col, 1:5))
}
}
