% Generated by tools/generate-man.R -- do not edit by hand
\name{cumulateseries}
\alias{cumulateseries}
\title{Cumulative Sum of a Time Series}
\description{
Wrapper around cumsum with validation.
}
\usage{
cumulateseries(seriesin)
}
\arguments{
  \item{seriesin}{Numeric vector. }
}
\value{
Numeric vector of cumulative sums. 
}
\seealso{

Other time-series: \code{\link{chgseries}}, \code{\link{compoundseries}}, \code{\link{lagdataframe}}, \code{\link{lagseries}}, \code{\link{leadlagpanel}}, \code{\link{leadlagseries}}, \code{\link{leadseries}}, \code{\link{ma}}, \code{\link{pchgseries}}
}
\examples{
\dontrun{
# Simple cumulative sum
iaw$cumulateseries(c(1, 2, 3, 4, 5))  # 1 3 6 10 15

# Cumulative returns: start at 0, accumulate daily P&L
daily_pnl <- c(0.01, -0.005, 0.02, -0.01, 0.015)
iaw$cumulateseries(daily_pnl)

# Negative steps produce a declining series
iaw$cumulateseries(c(10, -3, -2, -1, 5))
}
}
