% Generated by tools/generate-man.R -- do not edit by hand
\name{dependent.rank}
\alias{dependent.rank}
\title{Dependent Variable Rank}
\description{
Ranks within groups.
}
\usage{
dependent.rank(x, groups = NULL)
}
\arguments{
  \item{x}{Values to rank.}
  \item{groups}{Grouping variable. }
}
\value{
Vector of ranks. 
}
\seealso{

Other statistics: \code{\link{autocorrel}}, \code{\link{autopcorrel}}, \code{\link{cmpsum}}, \code{\link{colN}}, \code{\link{colSds}}, \code{\link{covp}}, \code{\link{print.cor}}, \code{\link{rank}}, \code{\link{rowSds}}, \code{\link{sdp}}, \code{\link{sse}}, \code{\link{sst}}, \code{\link{summary}}, \code{\link{tabularsummary}}, \code{\link{varp}}
}
