% Generated by tools/generate-man.R -- do not edit by hand
\name{dict.lookup}
\alias{dict.lookup}
\title{Dictionary Lookup}
\description{
NOTE: This was redesigned from a data-frame merge interface (see Rold/dict-lookup.R)
to a named-vector lookup. Verify this was intentional.
Looks up values in a dictionary.
}
\usage{
dict.lookup(keys, dict, default = NA)
}
\arguments{
  \item{keys}{Keys to look up.}
  \item{dict}{Named vector (dictionary).}
  \item{default}{Default for missing. }
}
\value{
Looked up values. 
}
\seealso{

Other utilities: \code{\link{ascii}}, \code{\link{base}}, \code{\link{beep}}, \code{\link{bisection}}, \code{\link{cat.both}}, \code{\link{cat.stderr}}, \code{\link{chatter}}, \code{\link{check.names}}, \code{\link{debug.advice}}, \code{\link{debug.on}}, \code{\link{done}}, \code{\link{enable.error.line.num}}, \code{\link{exit}}, \code{\link{grepcolname}}, \code{\link{heartbeat}}, \code{\link{index.of.variable}}, \code{\link{instrumentR}}, \code{\link{kable}}, \code{\link{latex}}, \code{\link{latlon.distance}}, \code{\link{latlonx}}, \code{\link{mklatlonid}}, \code{\link{invlatlon}}, \code{\link{testlatlon}}, \code{\link{ls.objects}}, \code{\link{meminfo}}, \code{\link{mk.random.walk}}, \code{\link{mkRdictionary}}, \code{\link{msg}}, \code{\link{nearest}}, \code{\link{now}}, \code{\link{numobs.by}}, \code{\link{object.size.MB}}, \code{\link{osinfo}}, \code{\link{rdraw.perfect}}, \code{\link{require.variables}}, \code{\link{rexp.perfect}}, \code{\link{rnorm.perfect}}, \code{\link{runall}}, \code{\link{setupknitr}}, \code{\link{sink}}, \code{\link{sprint.data.frame}}, \code{\link{strcat}}, \code{\link{whatis}}, \code{\link{which.nearest}}
}
\examples{
\dontrun{
d <- c(a = 1, b = 2, c = 3)
iaw$dict.lookup(c("a", "c", "x"), d, NA)
}
}
