% Generated by tools/generate-man.R -- do not edit by hand
\name{diffdays}
\alias{diffdays}
\title{Calculate Days Between Dates}
\description{
Calculates difference in days between dates.
}
\usage{
diffdays(d1, d2)
}
\arguments{
  \item{d1}{First date.}
  \item{d2}{Second date. }
}
\value{
Numeric days. 
}
\seealso{

Other datetime: \code{\link{epoch2nyc}}, \code{\link{epoch2POSIXct}}, \code{\link{epoch2yyyymmdd}}, \code{\link{tmdiffsec}}, \code{\link{yyyymmdd.toggle}}, \code{\link{yyyymmdd}}
}
\examples{
\dontrun{
iaw$diffdays(as.Date("2021-01-01"), as.Date("2021-01-15"))
}
}
