% Generated by tools/generate-man.R -- do not edit by hand
\name{epoch2nyc}
\alias{epoch2nyc}
\title{Convert Unix Epoch to NYC Time}
\description{
Converts Unix timestamp to NYC timezone.
}
\usage{
epoch2nyc(epoch)
}
\arguments{
  \item{epoch}{Numeric Unix timestamp. }
}
\value{
POSIXct in America/New_York. 
}
\seealso{

Other datetime: \code{\link{diffdays}}, \code{\link{epoch2POSIXct}}, \code{\link{epoch2yyyymmdd}}, \code{\link{tmdiffsec}}, \code{\link{yyyymmdd.toggle}}, \code{\link{yyyymmdd}}
}
\examples{
\dontrun{
# Unix epoch 0 is 1969-12-31 19:00:00 EST
iaw$epoch2nyc(0)

# Convert a known timestamp: 2021-01-01 00:00:00 UTC = 2020-12-31 19:00:00 EST
iaw$epoch2nyc(1609459200)

# Convert a vector of timestamps
ts <- c(1609459200, 1609545600)
iaw$epoch2nyc(ts)

}
}
