% Generated by tools/generate-man.R -- do not edit by hand
\name{is.character}
\alias{is.character}
\title{Test if Object is Character Vector of Specific Length}
\description{
Extends base is.character to also check length.
}
\usage{
is.character(cvec, required.length.of.cvec = NULL)
}
\arguments{
  \item{cvec}{Object to test.}
  \item{required.length.of.cvec}{Required length, or \code{NULL} to skip length check (default \code{NULL}). }
}
\value{
Logical scalar. 
}
\seealso{

Other type-checking: \code{\link{is.instring}}, \code{\link{is.numeric}}, \code{\link{is.scalar}}
}
\examples{
\dontrun{
iaw$is.character("hello", 1)
iaw$is.character(c("a", "b"), 2)
}
}
