% Generated by tools/generate-man.R -- do not edit by hand
\name{is.instring}
\alias{is.instring}
\title{Test if Pattern Exists in String}
\description{
A wrapper for grepl.
}
\usage{
is.instring(needle, heystack)
}
\arguments{
  \item{needle}{Character pattern to search for.}
  \item{heystack}{Character vector to search within. }
}
\value{
Logical vector. 
}
\seealso{

Other type-checking: \code{\link{is.character}}, \code{\link{is.numeric}}, \code{\link{is.scalar}}
}
\examples{
\dontrun{
iaw$is.instring("ab", c("abc", "def", "ab"))
}
}
