% Generated by tools/generate-man.R -- do not edit by hand
\name{kill.useless.cols}
\alias{kill.useless.cols}
\title{Remove Constant Columns}
\description{
Removes columns with no variance.
}
\usage{
kill.useless.cols(d)
}
\arguments{
  \item{d}{Data frame. }
}
\value{
Data frame with constant columns removed. 
}
\seealso{

Other data-manipulation: \code{\link{diffid.for.fixed.effects}}, \code{\link{recode}}, \code{\link{rename.columns}}, \code{\link{subset.data.frame}}
}
\examples{
\dontrun{
df <- data.frame(a = 1:3, b = c(1,1,1))
iaw$kill.useless.cols(df)
}
}
