% Generated by tools/generate-man.R -- do not edit by hand
\name{lagdataframe}
\alias{lagdataframe}
\title{Lag Data Frame Columns}
\description{
Creates lagged versions of data frame columns.
}
\usage{
lagdataframe(df, vars, nlags = 1, panelid = NULL, timeid = NULL)
}
\arguments{
  \item{df}{Data frame.}
  \item{vars}{Variables to lag.}
  \item{nlags}{Number of lags.}
  \item{panelid}{Panel identifier.}
  \item{timeid}{Time identifier. }
}
\value{
Data frame with lagged columns. 
}
\seealso{

Other time-series: \code{\link{chgseries}}, \code{\link{compoundseries}}, \code{\link{cumulateseries}}, \code{\link{lagseries}}, \code{\link{leadlagpanel}}, \code{\link{leadlagseries}}, \code{\link{leadseries}}, \code{\link{ma}}, \code{\link{pchgseries}}
}
