% Generated by tools/generate-man.R -- do not edit by hand
\name{lagseries}
\alias{lagseries}
\title{Lag a Time Series}
\description{
Shifts series backward by numlags periods.
}
\usage{
lagseries(series, numlags = 1, panelid = NULL, timeid = NULL)
}
\arguments{
  \item{series}{A numeric vector.}
  \item{numlags}{Integer number of lags. Default 1.}
  \item{panelid}{Optional panel identifier vector.}
  \item{timeid}{Optional time identifier vector. }
}
\value{
Numeric vector of same length. 
}
\seealso{

Other time-series: \code{\link{chgseries}}, \code{\link{compoundseries}}, \code{\link{cumulateseries}}, \code{\link{lagdataframe}}, \code{\link{leadlagpanel}}, \code{\link{leadlagseries}}, \code{\link{leadseries}}, \code{\link{ma}}, \code{\link{pchgseries}}
}
\examples{
\dontrun{
x <- c(1, 2, 3, 4, 5)
iaw$lagseries(x)
}
}
