% Generated by tools/generate-man.R -- do not edit by hand
\name{leadlagpanel}
\alias{leadlagpanel}
\title{Lead/Lag Panel Data}
\description{
Creates lead/lag variables for panel data.
}
\usage{
leadlagpanel(df, vars, nlags, panelid, timeid)
}
\arguments{
  \item{df}{Data frame.}
  \item{vars}{Variables to transform.}
  \item{nlags}{Number of lags (negative for leads).}
  \item{panelid}{Panel ID column name.}
  \item{timeid}{Time column name. }
}
\value{
Data frame with new columns. 
}
\seealso{

Other time-series: \code{\link{chgseries}}, \code{\link{compoundseries}}, \code{\link{cumulateseries}}, \code{\link{lagdataframe}}, \code{\link{lagseries}}, \code{\link{leadlagseries}}, \code{\link{leadseries}}, \code{\link{ma}}, \code{\link{pchgseries}}
}
