% Generated by tools/generate-man.R -- do not edit by hand
\name{lock}
\alias{lock}
\alias{unlock}
\title{File Locking}
\description{
Creates a lock file.
}
\usage{
lock(lockfile, timeout = 60)
unlock(lockfile)
}
\arguments{
  \item{lockfile}{Lock file path.}
  \item{timeout}{Timeout in seconds. }
}
\value{
TRUE if lock acquired. 
}
\seealso{

Other io: \code{\link{read.csv}}, \code{\link{.Riolog}}, \code{\link{touch}}, \code{\link{write.csv}}
}
