% Generated by tools/generate-man.R -- do not edit by hand
\name{ma}
\alias{ma}
\title{Moving Average}
\description{
Calculates moving average.
}
\usage{
ma(x, n, around = FALSE)
}
\arguments{
  \item{x}{Numeric vector.}
  \item{n}{Window size.}
  \item{around}{Logical. If \code{TRUE}, center-aligned; if \code{FALSE} (default), right-aligned. }
}
\value{
Numeric vector of moving averages. 
}
\seealso{

Other time-series: \code{\link{chgseries}}, \code{\link{compoundseries}}, \code{\link{cumulateseries}}, \code{\link{lagdataframe}}, \code{\link{lagseries}}, \code{\link{leadlagpanel}}, \code{\link{leadlagseries}}, \code{\link{leadseries}}, \code{\link{pchgseries}}
}
\examples{
\dontrun{
iaw$ma(1:10, 3)
}
}
