% Generated by tools/generate-man.R -- do not edit by hand
\name{mc.replicate}
\alias{mc.replicate}
\title{Parallel Replicate}
\description{
Parallel version of replicate.
}
\usage{
mc.replicate(n, expr, simplify = TRUE)
}
\arguments{
  \item{n}{Number of replications.}
  \item{expr}{Expression to evaluate.}
  \item{simplify}{Simplify result. }
}
\value{
Vector or list of results. 
}
\seealso{

Other parallel: \code{\link{mc.by}}, \code{\link{mclapply}}, \code{\link{mcsapply}}, \code{\link{oc.by}}, \code{\link{rbind.mc.by}}, \code{\link{rbind.oc.by}}
}
\examples{
\dontrun{
iaw$mc.replicate(100, mean(rnorm(100)))
}
}
