% Generated by tools/generate-man.R -- do not edit by hand
\name{mcsapply}
\alias{mcsapply}
\alias{mc.sapply}
\title{Enhanced Multicore Sapply}
\description{
Wrapper for mclapply with simplification.
}
\usage{
mcsapply(X, FUN, ...)
}
\arguments{
  \item{X}{List or vector.}
  \item{FUN}{Function to apply.}
  \item{...}{Arguments to FUN. }
}
\value{
Simplified result. 
}
\seealso{

Other parallel: \code{\link{mc.by}}, \code{\link{mc.replicate}}, \code{\link{mclapply}}, \code{\link{oc.by}}, \code{\link{rbind.mc.by}}, \code{\link{rbind.oc.by}}
}
\examples{
\dontrun{
iaw$mcsapply(1:10, function(x) x^2)
}
}
