% Generated by tools/generate-man.R -- do not edit by hand
\name{merge4seq}
\alias{merge4seq}
\title{Merge for Sequential Data}
\description{
Merges data frames maintaining order.
}
\usage{
merge4seq(x, y, by, ...)
}
\arguments{
  \item{x}{First data frame.}
  \item{y}{Second data frame.}
  \item{by}{Merge keys.}
  \item{...}{Additional merge arguments. }
}
\value{
Merged data frame. 
}
\seealso{

Other data-reshaping: \code{\link{long2wide}}, \code{\link{multimerge}}, \code{\link{wide2long}}
}
