% Generated by tools/generate-man.R -- do not edit by hand
\name{multimerge}
\alias{multimerge}
\title{Merge Multiple Data Frames}
\description{
Merges list of data frames.
}
\usage{
multimerge(dflist, by, ...)
}
\arguments{
  \item{dflist}{List of data frames.}
  \item{by}{Merge keys.}
  \item{...}{Additional merge arguments. }
}
\value{
Merged data frame. 
}
\seealso{

Other data-reshaping: \code{\link{long2wide}}, \code{\link{merge4seq}}, \code{\link{wide2long}}
}
