% Generated by tools/generate-man.R -- do not edit by hand
\name{oc.by}
\alias{oc.by}
\title{One-Core By (Serial)}
\description{
Serial version of mc.by.
}
\usage{
oc.by(indata, INDICES, FUNIN, dataframeout=TRUE, ...)
}
\arguments{
  \item{indata}{Data frame.}
  \item{INDICES}{Grouping variable.}
  \item{FUNIN}{Function to apply.}
  \item{dataframeout}{helps with conversion reversion to original behavior}
  \item{...}{Additional arguments. }
}
\value{
List of results. 
}
\seealso{

Other parallel: \code{\link{mc.by}}, \code{\link{mc.replicate}}, \code{\link{mclapply}}, \code{\link{mcsapply}}, \code{\link{rbind.mc.by}}, \code{\link{rbind.oc.by}}
}
