% Generated by tools/generate-man.R -- do not edit by hand
\name{oc.by}
\alias{oc.by}
\title{One-Core By (Serial)}
\description{
Serial version of mc.by.
}
\usage{
oc.by(indata, INDICES, FUNIN, dataframeout=TRUE, ...)
}
\arguments{
  \item{indata}{Data frame.}
  \item{INDICES}{Grouping variable.}
  \item{FUNIN}{Function to apply.}
  \item{dataframeout}{helps with conversion reversion to original behavior}
  \item{...}{Additional arguments. }
}
\value{
List of results. 
}
\seealso{

Other parallel: \code{\link{mc.by}}, \code{\link{mc.replicate}}, \code{\link{mclapply}}, \code{\link{mcsapply}}, \code{\link{rbind.mc.by}}, \code{\link{rbind.oc.by}}
}
\examples{
\dontrun{
# Compute per-group column means (serial version of mc.by)
df <- data.frame(
  group = c("A", "A", "B", "B", "B"),
  x     = c(1, 3, 2, 4, 6),
  y     = c(10, 20, 30, 40, 50)
)
result <- iaw$oc.by(df, df$group, function(sub) colMeans(sub[, c("x", "y")]))
do.call(rbind, result)

# Add a derived column within each group
iaw$oc.by(df, df$group, function(sub) {
  sub$z <- sub$x - mean(sub$x)
  sub
})

}
}
