% Generated by tools/generate-man.R -- do not edit by hand
\name{plot.boxtext}
\alias{plot.boxtext}
\title{Text with Box Background}
\description{
Adds text with background box.
}
\usage{
\method{plot}{boxtext}(x, y, labels, ...)
}
\arguments{
  \item{x}{X position.}
  \item{y}{Y position.}
  \item{labels}{Text labels.}
  \item{...}{Text arguments. }
}
\value{
Invisible NULL. 
}
\seealso{

Other plotting: \code{\link{as.PlotDate.yyyymmdd}}, \code{\link{colorblind}}, \code{\link{plot.contour}}, \code{\link{plot.ecdf.lineplot}}, \code{\link{plot.hist2d}}, \code{\link{hline}}, \code{\link{iaw.boxplot}}, \code{\link{plot.labelcurve}}, \code{\link{lambda2color}}, \code{\link{plot.loess.lineplot}}, \code{\link{MixColor}}, \code{\link{plot.native.slope}}, \code{\link{plot.p.arrows}}, \code{\link{plot.pdf.embedfonts}}, \code{\link{pdf.end}}, \code{\link{plot.pdf.start.bera}}, \code{\link{pdf.start}}, \code{\link{plot.timearea}}, \code{\link{plot.worldmap}}, \code{\link{plot.points.toomany}}, \code{\link{plot.shade.xyy}}, \code{\link{plot.text.bg}}, \code{\link{plot.text.onbg}}, \code{\link{vline}}, \code{\link{plot.yearly.tickmarks}}, \code{\link{plot.ylab.left}}
}
\examples{
\dontrun{
# Add a label with a box background on an existing plot
plot(1:10, 1:10)
iaw$plot.boxtext(5, 5, "Center label")

# Customize text appearance
plot(1:10, 1:10, type = "n")
iaw$plot.boxtext(3, 7, "Group A", col = "blue", cex = 1.2)
iaw$plot.boxtext(7, 3, "Group B", col = "red",  cex = 1.2)

# Annotate key data points on a time-series plot
x <- 1:20
y <- cumsum(rnorm(20))
plot(x, y, type = "l", main = "Annotated series")
peak <- which.max(y)
iaw$plot.boxtext(peak, y[peak], "Peak", pos = 3, cex = 0.8)

# Label regression line at midpoint
plot(mtcars$wt, mtcars$mpg, pch = 16)
abline(lm(mpg ~ wt, data = mtcars), col = "red")
iaw$plot.boxtext(3.5, 22, "OLS fit", col = "red", font = 2)
}

}
