% Generated by tools/generate-man.R -- do not edit by hand
\name{plot.contour}
\alias{plot.contour}
\title{Contour Plot}
\description{
Creates filled contour plot.
}
\usage{
plot.contour(x, y, z, ...)
}
\arguments{
  \item{x}{X coordinates.}
  \item{y}{Y coordinates.}
  \item{z}{Matrix of z values.}
  \item{...}{Additional arguments. }
}
\value{
Invisible NULL. 
}
\seealso{

Other plotting: \code{\link{as.PlotDate.yyyymmdd}}, \code{\link{plot.boxtext}}, \code{\link{colorblind}}, \code{\link{plot.ecdf.lineplot}}, \code{\link{plot.hist2d}}, \code{\link{hline}}, \code{\link{iaw.boxplot}}, \code{\link{plot.labelcurve}}, \code{\link{lambda2color}}, \code{\link{plot.loess.lineplot}}, \code{\link{MixColor}}, \code{\link{plot.native.slope}}, \code{\link{plot.p.arrows}}, \code{\link{plot.pdf.embedfonts}}, \code{\link{pdf.end}}, \code{\link{pdf.start}}, \code{\link{plot.pdf.start.bera}}, \code{\link{plot.timearea}}, \code{\link{plot.worldmap}}, \code{\link{plot.points.toomany}}, \code{\link{plot.shade.xyy}}, \code{\link{plot.text.bg}}, \code{\link{plot.text.onbg}}, \code{\link{vline}}, \code{\link{plot.yearly.tickmarks}}, \code{\link{plot.ylab.left}}
}
\examples{
\dontrun{
# Simple filled contour of a bivariate normal density surface
x <- seq(-3, 3, length.out = 30)
y <- seq(-3, 3, length.out = 30)
z <- outer(x, y, function(a, b) exp(-0.5 * (a^2 + b^2)))
iaw$plot.contour(x, y, z)

# Add a color palette and axis labels
iaw$plot.contour(x, y, z,
                 color.palette = heat.colors,
                 xlab = "X", ylab = "Y",
                 main = "Bivariate density")
}

}
