% Generated by tools/generate-man.R -- do not edit by hand
\name{plot.labelcurve}
\alias{plot.labelcurve}
\title{Place a text label on a line, based on slope at an x-point}
\description{
places text label on slope at an x-point
}
\usage{
plot.labelcurve(xseries, yseries, xval, label, ...)
}
\arguments{
  \item{xseries}{X values.}
  \item{yseries}{Y values.}
  \item{xval}{the x-location for the label}
  \item{label}{the text content of the label}
  \item{...}{Plot arguments. }
}
\value{
Invisible NULL. 
}
\seealso{

Other plotting: \code{\link{as.PlotDate.yyyymmdd}}, \code{\link{plot.boxtext}}, \code{\link{colorblind}}, \code{\link{plot.contour}}, \code{\link{plot.ecdf.lineplot}}, \code{\link{plot.hist2d}}, \code{\link{hline}}, \code{\link{iaw.boxplot}}, \code{\link{lambda2color}}, \code{\link{plot.loess.lineplot}}, \code{\link{MixColor}}, \code{\link{plot.native.slope}}, \code{\link{plot.p.arrows}}, \code{\link{plot.pdf.embedfonts}}, \code{\link{pdf.end}}, \code{\link{plot.pdf.start.bera}}, \code{\link{pdf.start}}, \code{\link{plot.timearea}}, \code{\link{plot.worldmap}}, \code{\link{plot.points.toomany}}, \code{\link{plot.shade.xyy}}, \code{\link{plot.text.bg}}, \code{\link{plot.text.onbg}}, \code{\link{vline}}, \code{\link{plot.yearly.tickmarks}}, \code{\link{plot.ylab.left}}
}
