% Generated by tools/generate-man.R -- do not edit by hand
\name{plot.loess.lineplot}
\alias{plot.loess.lineplot}
\title{Loess Smoothed Line Plot}
\description{
Plots loess-smoothed line.
}
\usage{
plot.loess.lineplot(x, y, lo=200, span=0.75, mean.hatch=NA, ...)
}
\arguments{
  \item{x}{Numeric vector of x values.}
  \item{y}{Numeric vector of y values.}
  \item{lo}{Number of points for the smoothed line (default 200).}
  \item{span}{Smoothing span for \code{loess} (default 0.75).}
  \item{mean.hatch}{X-value at which to draw a cross-hatch mark, or \code{NA} to skip (default \code{NA}).}
  \item{...}{Additional arguments passed to \code{lines}. }
}
\value{
Data frame of smoothed x and y values (invisibly). 
}
\seealso{

Other plotting: \code{\link{as.PlotDate.yyyymmdd}}, \code{\link{plot.boxtext}}, \code{\link{colorblind}}, \code{\link{plot.contour}}, \code{\link{plot.ecdf.lineplot}}, \code{\link{plot.hist2d}}, \code{\link{hline}}, \code{\link{iaw.boxplot}}, \code{\link{plot.labelcurve}}, \code{\link{lambda2color}}, \code{\link{MixColor}}, \code{\link{plot.native.slope}}, \code{\link{plot.p.arrows}}, \code{\link{plot.pdf.embedfonts}}, \code{\link{pdf.end}}, \code{\link{pdf.start}}, \code{\link{plot.pdf.start.bera}}, \code{\link{plot.timearea}}, \code{\link{plot.worldmap}}, \code{\link{plot.points.toomany}}, \code{\link{plot.shade.xyy}}, \code{\link{plot.text.bg}}, \code{\link{plot.text.onbg}}, \code{\link{vline}}, \code{\link{plot.yearly.tickmarks}}, \code{\link{plot.ylab.left}}
}
\examples{
\dontrun{
# Smooth a noisy sine wave and overlay on scatter plot
set.seed(1)
x <- runif(100, 0, 2 * pi)
y <- sin(x) + rnorm(100, sd = 0.3)
plot(x, y, pch = 16, col = "gray70", main = "Loess smoother")
iaw$plot.loess.lineplot(x, y, col = "blue", lwd = 2)

# Add a cross-hatch mark at the mean of x
plot(x, y, pch = 16, col = "gray70")
iaw$plot.loess.lineplot(x, y, mean.hatch = pi, col = "red", lwd = 2)

# Tighter span for less smoothing
plot(x, y, pch = 16, col = "gray70")
iaw$plot.loess.lineplot(x, y, span = 0.3, col = "darkgreen", lwd = 2)
}

}
