% Generated by tools/generate-man.R -- do not edit by hand
\name{plot.text.onbg}
\alias{plot.text.onbg}
\title{Text on Background}
\description{
Alias for text with background.
}
\arguments{
  \item{x}{X position.}
  \item{y}{Y position.}
  \item{labels}{Text.}
  \item{...}{Arguments. }
}
\value{
Invisible NULL. 
}
\seealso{

Other plotting: \code{\link{as.PlotDate.yyyymmdd}}, \code{\link{plot.boxtext}}, \code{\link{colorblind}}, \code{\link{plot.contour}}, \code{\link{plot.ecdf.lineplot}}, \code{\link{plot.hist2d}}, \code{\link{hline}}, \code{\link{iaw.boxplot}}, \code{\link{plot.labelcurve}}, \code{\link{lambda2color}}, \code{\link{plot.loess.lineplot}}, \code{\link{MixColor}}, \code{\link{plot.native.slope}}, \code{\link{plot.p.arrows}}, \code{\link{plot.pdf.embedfonts}}, \code{\link{pdf.end}}, \code{\link{pdf.start}}, \code{\link{plot.pdf.start.bera}}, \code{\link{plot.timearea}}, \code{\link{plot.worldmap}}, \code{\link{plot.points.toomany}}, \code{\link{plot.shade.xyy}}, \code{\link{plot.text.bg}}, \code{\link{vline}}, \code{\link{plot.yearly.tickmarks}}, \code{\link{plot.ylab.left}}
}
\examples{
\dontrun{
# Alias for iaw$plot.text.bg - adds text with a colored background
plot(1:10, 1:10, type = "n")
iaw$plot.text.onbg(5, 5, "center", bg = "lightyellow")
}

}
\keyword{internal}
