% Generated by tools/generate-man.R -- do not edit by hand
\name{plot.timearea}
\alias{plot.timearea}
\title{Time Series Area Plot}
\description{
Creates time series area plot.
}
\usage{
plot.timearea(x, y, ...)
}
\arguments{
  \item{x}{Time values.}
  \item{y}{Y values.}
  \item{...}{Polygon arguments. }
}
\value{
Invisible NULL. 
}
\seealso{

Other plotting: \code{\link{as.PlotDate.yyyymmdd}}, \code{\link{plot.boxtext}}, \code{\link{colorblind}}, \code{\link{plot.contour}}, \code{\link{plot.ecdf.lineplot}}, \code{\link{plot.hist2d}}, \code{\link{hline}}, \code{\link{iaw.boxplot}}, \code{\link{plot.labelcurve}}, \code{\link{lambda2color}}, \code{\link{plot.loess.lineplot}}, \code{\link{MixColor}}, \code{\link{plot.native.slope}}, \code{\link{plot.p.arrows}}, \code{\link{plot.pdf.embedfonts}}, \code{\link{pdf.end}}, \code{\link{pdf.start}}, \code{\link{plot.pdf.start.bera}}, \code{\link{plot.worldmap}}, \code{\link{plot.points.toomany}}, \code{\link{plot.shade.xyy}}, \code{\link{plot.text.bg}}, \code{\link{plot.text.onbg}}, \code{\link{vline}}, \code{\link{plot.yearly.tickmarks}}, \code{\link{plot.ylab.left}}
}
\examples{
\dontrun{
# Area chart for a simple time series
t <- 1:24
y <- c(2, 3, 4, 3, 5, 6, 5, 4, 6, 8, 7, 6,
       5, 4, 5, 6, 7, 8, 9, 8, 7, 6, 5, 4)
iaw$plot.timearea(t, y, col = "steelblue", main = "Monthly values")

# Highlight positive area in green, overlap with a line
iaw$plot.timearea(t, y, col = adjustcolor("green", alpha.f = 0.4))
lines(t, y, lwd = 2)
}

}
