% Generated by tools/generate-man.R -- do not edit by hand
\name{plot.worldmap}
\alias{plot.worldmap}
\title{World Map Plot}
\description{
Creates basic world map.
}
\usage{
\method{plot}{worldmap}(...)
}
\arguments{
  \item{...}{Map arguments. }
}
\value{
Invisible NULL. 
}
\seealso{

Other plotting: \code{\link{as.PlotDate.yyyymmdd}}, \code{\link{plot.boxtext}}, \code{\link{colorblind}}, \code{\link{plot.contour}}, \code{\link{plot.ecdf.lineplot}}, \code{\link{plot.hist2d}}, \code{\link{hline}}, \code{\link{iaw.boxplot}}, \code{\link{plot.labelcurve}}, \code{\link{lambda2color}}, \code{\link{plot.loess.lineplot}}, \code{\link{MixColor}}, \code{\link{plot.native.slope}}, \code{\link{plot.p.arrows}}, \code{\link{plot.pdf.embedfonts}}, \code{\link{pdf.end}}, \code{\link{pdf.start}}, \code{\link{plot.pdf.start.bera}}, \code{\link{plot.timearea}}, \code{\link{plot.points.toomany}}, \code{\link{plot.shade.xyy}}, \code{\link{plot.text.bg}}, \code{\link{plot.text.onbg}}, \code{\link{vline}}, \code{\link{plot.yearly.tickmarks}}, \code{\link{plot.ylab.left}}
}
\examples{
\dontrun{
# Draw a basic world map outline
iaw$plot.worldmap()

# Draw with fill and highlight a region
iaw$plot.worldmap(fill = TRUE, col = "lightgray")
maps::map("world", regions = "USA", fill = TRUE, col = "steelblue", add = TRUE)

# Restrict to Europe
iaw$plot.worldmap(xlim = c(-10, 40), ylim = c(35, 70),
                  fill = TRUE, col = "lightyellow")
}

}
