% Generated by tools/generate-man.R -- do not edit by hand
\name{rbind.mc.by}
\alias{rbind.mc.by}
\title{Rbind Results from mc.by}
\description{
Combines mc.by results into data frame.
}
\usage{
rbind.mc.by(indata, INDICES, FUNIN, ...)
}
\arguments{
  \item{indata}{Data frame.}
  \item{INDICES}{Grouping variable.}
  \item{FUNIN}{Function returning data frames.}
  \item{...}{Additional arguments. }
}
\value{
Combined data frame. 
}
\seealso{

Other parallel: \code{\link{mc.by}}, \code{\link{mc.replicate}}, \code{\link{mclapply}}, \code{\link{mcsapply}}, \code{\link{oc.by}}, \code{\link{rbind.oc.by}}
}
