% Generated by tools/generate-man.R -- do not edit by hand
\name{rbind.oc.by}
\alias{rbind.oc.by}
\title{Rbind Results from oc.by}
\description{
Combines oc.by results into data frame.
}
\usage{
rbind.oc.by(indata, INDICES, FUNIN, ...)
}
\arguments{
  \item{indata}{Data frame.}
  \item{INDICES}{Grouping variable.}
  \item{FUNIN}{Function returning data frames.}
  \item{...}{Additional arguments. }
}
\value{
Combined data frame. 
}
\seealso{

Other parallel: \code{\link{mc.by}}, \code{\link{mc.replicate}}, \code{\link{mclapply}}, \code{\link{mcsapply}}, \code{\link{oc.by}}, \code{\link{rbind.mc.by}}
}
\examples{
\dontrun{
# Demean x within each group and rbind all groups back together
df <- data.frame(
  group = c("A", "A", "B", "B", "B"),
  x     = c(1, 3, 2, 4, 6)
)
result <- iaw$rbind.oc.by(df, df$group, function(sub) {
  sub$x_dm <- sub$x - mean(sub$x)
  sub
})
result  # 5-row data frame with x_dm column added

# Filter rows within groups and recombine
iaw$rbind.oc.by(df, df$group, function(sub) sub[sub$x > mean(sub$x), ])

}
}
