% Generated by tools/generate-man.R -- do not edit by hand
\name{rename.columns}
\alias{rename.columns}
\alias{rename.column}
\title{Rename Columns in Data Frame}
\description{
Renames columns using from/to mapping or named vector.
}
\usage{
rename.columns(df, from, to = NULL)
}
\arguments{
  \item{df}{Data frame or list.}
  \item{from}{Current names or named vector (old=new).}
  \item{to}{New names (ignored if from is named). }
}
\value{
Modified data frame or list. 
}
\seealso{

Other data-manipulation: \code{\link{diffid.for.fixed.effects}}, \code{\link{kill.useless.cols}}, \code{\link{recode}}, \code{\link{subset.data.frame}}
}
\examples{
\dontrun{
df <- data.frame(a = 1, b = 2)
iaw$rename.columns(df, c(a = "x", b = "y"))
}
}
