% Generated by tools/generate-man.R -- do not edit by hand
\name{rowSds}
\alias{rowSds}
\title{Row Standard Deviations}
\description{
Calculates SD for each row.
}
\usage{
rowSds(x, na.rm = TRUE)
}
\arguments{
  \item{x}{Matrix or data frame.}
  \item{na.rm}{Remove NA values. }
}
\value{
Vector of standard deviations. 
}
\seealso{

Other statistics: \code{\link{autocorrel}}, \code{\link{autopcorrel}}, \code{\link{cmpsum}}, \code{\link{colN}}, \code{\link{colSds}}, \code{\link{covp}}, \code{\link{dependent.rank}}, \code{\link{print.cor}}, \code{\link{rank}}, \code{\link{sdp}}, \code{\link{sse}}, \code{\link{sst}}, \code{\link{summary}}, \code{\link{tabularsummary}}, \code{\link{varp}}
}
\examples{
\dontrun{
iaw$rowSds(matrix(rnorm(20), nrow = 4))
}
}
