% Generated by tools/generate-man.R -- do not edit by hand
\name{sdp}
\alias{sdp}
\title{Population Standard Deviation}
\description{
Calculates population SD (n denominator).
}
\usage{
sdp(x, na.rm = TRUE)
}
\arguments{
  \item{x}{Numeric vector.}
  \item{na.rm}{Remove NA values. }
}
\value{
Standard deviation value. 
}
\seealso{

Other statistics: \code{\link{autocorrel}}, \code{\link{autopcorrel}}, \code{\link{cmpsum}}, \code{\link{colN}}, \code{\link{colSds}}, \code{\link{covp}}, \code{\link{dependent.rank}}, \code{\link{print.cor}}, \code{\link{rank}}, \code{\link{rowSds}}, \code{\link{sse}}, \code{\link{sst}}, \code{\link{summary}}, \code{\link{tabularsummary}}, \code{\link{varp}}
}
\examples{
\dontrun{
iaw$sdp(1:10)
}
}
