% Generated by tools/generate-man.R -- do not edit by hand
\name{sdp}
\alias{sdp}
\title{Population Standard Deviation}
\description{
Calculates population SD (n denominator).
}
\usage{
sdp(x, na.rm = TRUE)
}
\arguments{
  \item{x}{Numeric vector.}
  \item{na.rm}{Remove NA values. }
}
\value{
Standard deviation value. 
}
\seealso{

Other statistics: \code{\link{autocorrel}}, \code{\link{autopcorrel}}, \code{\link{cmpsum}}, \code{\link{colN}}, \code{\link{colSds}}, \code{\link{covp}}, \code{\link{dependent.rank}}, \code{\link{print.cor}}, \code{\link{rank}}, \code{\link{rowSds}}, \code{\link{sse}}, \code{\link{sst}}, \code{\link{summary}}, \code{\link{tabularsummary}}, \code{\link{varp}}
}
\examples{
\dontrun{
iaw$sdp(1:10)          # population SD ~2.872 (vs sd() ~3.028)

# Compare with sample SD: sdp is always smaller
x <- c(2, 4, 4, 4, 5, 5, 7, 9)
iaw$sdp(x)             # 2 (exact population SD of this population)
sd(x)                  # 2.138 (sample SD with n-1)

# NA values are removed by default
iaw$sdp(c(1, 2, NA, 4, 5))

# Constant vector: SD is 0
iaw$sdp(rep(5, 10))    # 0
}
}
