% Generated by tools/generate-man.R -- do not edit by hand
\name{sse}
\alias{sse}
\title{Sum of Squared Errors}
\description{
Calculates sum of squared errors.
}
\usage{
sse(actual, predicted)
}
\arguments{
  \item{actual}{Actual values.}
  \item{predicted}{Predicted values. }
}
\value{
SSE value. 
}
\seealso{

Other statistics: \code{\link{autocorrel}}, \code{\link{autopcorrel}}, \code{\link{cmpsum}}, \code{\link{colN}}, \code{\link{colSds}}, \code{\link{covp}}, \code{\link{dependent.rank}}, \code{\link{print.cor}}, \code{\link{rank}}, \code{\link{rowSds}}, \code{\link{sdp}}, \code{\link{sst}}, \code{\link{summary}}, \code{\link{tabularsummary}}, \code{\link{varp}}
}
