% Generated by tools/generate-man.R -- do not edit by hand
\name{sst}
\alias{sst}
\title{Total Sum of Squares}
\description{
Calculates total sum of squares.
}
\usage{
sst(x)
}
\arguments{
  \item{x}{Numeric vector. }
}
\value{
SST value. 
}
\seealso{

Other statistics: \code{\link{autocorrel}}, \code{\link{autopcorrel}}, \code{\link{cmpsum}}, \code{\link{colN}}, \code{\link{colSds}}, \code{\link{covp}}, \code{\link{dependent.rank}}, \code{\link{print.cor}}, \code{\link{rank}}, \code{\link{rowSds}}, \code{\link{sdp}}, \code{\link{sse}}, \code{\link{summary}}, \code{\link{tabularsummary}}, \code{\link{varp}}
}
\examples{
\dontrun{
iaw$sst(c(1, 2, 3, 4, 5))
}
}
