% Generated by tools/generate-man.R -- do not edit by hand
\name{tmdiffsec}
\alias{tmdiffsec}
\title{Time Difference in Seconds}
\description{
Calculates time difference in seconds.
}
\usage{
tmdiffsec(t1, t2)
}
\arguments{
  \item{t1}{First time.}
  \item{t2}{Second time. }
}
\value{
Numeric seconds. 
}
\seealso{

Other datetime: \code{\link{diffdays}}, \code{\link{epoch2nyc}}, \code{\link{epoch2POSIXct}}, \code{\link{epoch2yyyymmdd}}, \code{\link{yyyymmdd.toggle}}, \code{\link{yyyymmdd}}
}
