% Generated by tools/generate-man.R -- do not edit by hand
\name{winsorize.level}
\alias{winsorize.level}
\title{Winsorize at Fixed Levels}
\description{
Clips values to specified min/max.
}
\usage{
winsorize.level(x, lvlminmax, name = NULL, verbose = FALSE)
}
\arguments{
  \item{x}{Numeric vector.}
  \item{lvlminmax}{Numeric vector of length 2: c(min, max).}
  \item{name}{Variable name for verbose output.}
  \item{verbose}{Print summary. }
}
\value{
Numeric vector. 
}
\seealso{

Other data-transformation: \code{\link{normalize}}, \code{\link{standardize}}, \code{\link{winsorize.percentile}}
}
\examples{
\dontrun{
x <- c(-100, 1, 2, 3, 100)
iaw$winsorize.level(x, c(0, 10))
}
}
